refactor(frontend): split api client and domain types into modular files
