+++ packages/frontend/src/lib/api.ts
@@ -1,3749 +1,9 @@
-import { formatNumber, formatPoints } from './format';
-import { isOAuthPlaceholderUrl } from '@plexus/shared';
-import { normalizeApiAccessList } from './apiFormats';
-import { dedupeAliasTargets, dedupeById, dedupeModels, dedupeStrings } from './modelOptions';
-
-import type { QuotaCheckerInfo } from '../types/quota';
-
-const API_BASE = ''; // Proxied via server.ts
-
-/**
- * Extract supported API types from the provider configuration.
- * Infers types from api_base_url field: if it's a record/map, the keys are the supported types.
- * If it's a string, we infer the type from the URL pattern.
- * @param apiBaseUrl The api_base_url from provider configuration
- * @returns Array of supported API types (e.g., ["chat"], ["messages"], ["chat", "messages"])
- */
-function inferProviderTypes(apiBaseUrl?: string | Record<string, string>): string[] {
-  if (!apiBaseUrl) {
-    return ['chat']; // Default fallback
-  }
-
-  if (typeof apiBaseUrl === 'string') {
-    // Single URL - infer type from URL pattern
-    const url = apiBaseUrl.trim().toLowerCase();
-
-    if (isOAuthPlaceholderUrl(apiBaseUrl)) {
-      return ['oauth'];
-    }
-
-    // Check for known patterns
-    if (url.includes('anthropic.com')) {
-      return ['messages'];
-    } else if (url.includes('generativelanguage.googleapis.com')) {
-      return ['gemini'];
-    } else {
-      // Default to 'chat' for OpenAI-compatible APIs
-      return ['chat'];
-    }
-  } else {
-    // Record/map format - keys are the supported types
-    return Object.keys(apiBaseUrl).filter((key) => {
-      const value = apiBaseUrl[key];
-      return typeof value === 'string' && value.length > 0;
-    });
-  }
-}
-
-/**
- * Shape of the principal returned by GET /v0/management/auth/verify.
- * Admins get just { role: 'admin' }; api-key users get the key metadata so
- * the frontend can render a scoped view without a follow-up call.
- */
-export type Principal =
-  | { role: 'admin' }
-  | {
-      role: 'limited';
-      keyName: string;
-      allowedProviders: string[];
-      allowedModels: string[];
-      excludedProviders: string[];
-      excludedModels: string[];
-      quotaName?: string | null;
-      comment?: string | null;
-    };
-
-/**
- * Verify a credential against the backend. Returns the resolved principal on
- * success, or null on 401/network error.
- */
-export async function verifyAdminKey(key: string): Promise<Principal | null> {
-  try {
-    const res = await fetch('/v0/management/auth/verify', {
-      method: 'GET',
-      headers: { 'x-admin-key': key },
-    });
-    if (res.status !== 200) return null;
-    const body = (await res.json()) as {
-      ok: boolean;
-      role: 'admin' | 'limited';
-      keyName?: string;
-      allowedProviders?: string[];
-      allowedModels?: string[];
-      excludedProviders?: string[];
-      excludedModels?: string[];
-      quotaName?: string | null;
-      comment?: string | null;
-    };
-    if (!body.ok) return null;
-    if (body.role === 'admin') return { role: 'admin' };
-    return {
-      role: 'limited',
-      keyName: body.keyName!,
-      allowedProviders: body.allowedProviders ?? [],
-      allowedModels: body.allowedModels ?? [],
-      excludedProviders: body.excludedProviders ?? [],
-      excludedModels: body.excludedModels ?? [],
-      quotaName: body.quotaName ?? null,
-      comment: body.comment ?? null,
-    };
-  } catch {
-    return null;
-  }
-}
-
-const fetchWithAuth = async (url: string, options: RequestInit = {}) => {
-  const headers = new Headers(options.headers || {});
-  const adminKey = localStorage.getItem('plexus_admin_key');
-  if (adminKey) {
-    headers.set('x-admin-key', adminKey);
-  }
-
-  const res = await fetch(url, { ...options, headers });
-
-  if (res.status === 401) {
-    // If unauthorized, clear key to trigger re-login
-    localStorage.removeItem('plexus_admin_key');
-    // Optional: Dispatch event or reload.
-    // Usually the React Context will catch this on next refresh, or we can reload here.
-    if (window.location.pathname !== '/ui/login') {
-      window.location.href = '/ui/login';
-    }
-  }
-  return res;
-};
-
-function normalizeQuotaCheckerInfo(checker: QuotaCheckerInfo): QuotaCheckerInfo {
-  return {
-    ...checker,
-    meters: Array.isArray(checker.meters) ? checker.meters : [],
-  };
-}
-
-export interface Stat {
-  label: string;
-  value: string | number;
-  change?: number;
-  icon?: string;
-}
-
-export interface UsageData {
-  timestamp: string;
-  requests: number;
-  tokens: number;
-  inputTokens: number;
-  outputTokens: number;
-  cachedTokens: number;
-  cacheWriteTokens: number;
-}
-
-export interface TodayMetrics {
-  requests: number;
-  inputTokens: number;
-  outputTokens: number;
-  reasoningTokens: number;
-  cachedTokens: number;
-  cacheWriteTokens: number;
-  totalCost: number;
-}
-
-/**
- * Represents one concurrency point returned by the backend.
- *
- * Semantics depend on query mode:
- * - mode='timeline': count is bucketed per provider+model in 1-minute buckets.
- * - mode='live': count is a current in-flight snapshot per provider+model.
- *
- * Used by the Live Metrics concurrency card and usage analytics views.
- */
-export interface ConcurrencyData {
-  /** The LLM provider name, e.g., "anthropic", "openai", "google" */
-  provider: string;
-  /** The canonical model name as resolved by the router, e.g., "claude-sonnet-4-20250514" */
-  model: string;
-  /** Number of requests that started within this 1-minute bucket */
-  count: number;
-  /** Start of the 1-minute bucket as epoch milliseconds (floored to nearest 60000ms) */
-  timestamp: number;
-}
-
-export interface DashboardData {
-  stats: Stat[];
-  usageData: UsageData[];
-  cooldowns: Cooldown[];
-  todayMetrics: TodayMetrics;
-}
-
-export interface PieChartDataPoint {
-  name: string;
-  requests: number;
-  tokens: number;
-  [key: string]: string | number; // Index signature for recharts compatibility
-}
-
-export interface ProviderPerformanceData {
-  provider: string;
-  model: string;
-  target_model?: string;
-  avg_ttft_ms: number;
-  min_ttft_ms: number;
-  max_ttft_ms: number;
-  avg_tokens_per_sec: number;
-  min_tokens_per_sec: number;
-  max_tokens_per_sec: number;
-  avg_e2e_tokens_per_sec: number;
-  min_e2e_tokens_per_sec: number;
-  max_e2e_tokens_per_sec: number;
-  sample_count: number;
-  last_updated: number;
-}
-
-export interface Provider {
-  id: string;
-  name: string;
-  type: string | string[];
-  apiBaseUrl?: string | Record<string, string>;
-  apiKey: string;
-  oauthProvider?: string;
-  oauthAccount?: string;
-  enabled: boolean;
-  disableCooldown?: boolean;
-  stallCooldown?: boolean;
-  allow100PercentUtilization?: boolean;
-  estimateTokens?: boolean;
-  useClaudeMasking?: boolean;
-  geminiThinkingEnabled?: boolean;
-  discount?: number;
-  headers?: Record<string, string>;
-  extraBody?: Record<string, any>;
-  models?: string[] | Record<string, any>;
-  quotaChecker?: {
-    type?: string;
-    enabled: boolean;
-    intervalMinutes: number;
-    options?: Record<string, unknown>;
-  };
-  modelAutosync?: {
-    enabled: boolean;
-    intervalMinutes: number;
-  };
-  adapter?: any[];
-  timeoutMs?: number;
-  maxConcurrency?: number | null;
-  auto_compat?: boolean;
-  // Per-provider stall detection overrides
-  stallTtfbMs?: number | null;
-  stallTtfbBytes?: number | null;
-  stallMinBps?: number | null;
-  stallWindowMs?: number | null;
-  stallGracePeriodMs?: number | null;
-  pi_ai_provider?: string;
-  compaction?: CompactionSettings;
-  rawPassthrough?: {
-    enabled: boolean;
-    baseUrl: string;
-    auth: 'bearer' | 'x-api-key' | 'x-goog-api-key';
-  };
-}
-
-export type McpServer = RemoteMcpServer | LocalMcpServer;
-
-export interface RemoteMcpServer {
-  mode?: 'remote_http';
-  upstream_url: string;
-  enabled: boolean;
-  headers?: Record<string, string>;
-  auth_scheme?: string | null;
-  rate_limit_cooldown_ms?: number;
-  quota_cooldown_ms?: number;
-}
-
-export interface LocalMcpServer {
-  mode: 'local_http';
-  enabled: boolean;
-  launcher: 'bunx' | 'uvx';
-  package: string;
-  args?: string[];
-  env?: Record<string, string>;
-  port: number;
-  path?: string;
-  startup_timeout_ms?: number;
-  headers?: Record<string, string>;
-  auth_scheme?: string | null;
-  rate_limit_cooldown_ms?: number;
-  quota_cooldown_ms?: number;
-}
-
-export interface McpServerKey {
-  id: number;
-  key: string;
-  is_active: boolean;
-  cooldown_until: string | null;
-}
-
-export interface LocalMcpRuntimeStatus {
-  serverName: string;
-  status: 'stopped' | 'starting' | 'running' | 'failed';
-  pid: number | null;
-  url: string | null;
-  lastError: string | null;
-  startedAt: string | null;
-  exitedAt: string | null;
-}
-
-export interface McpLogRecord {
-  request_id: string;
-  created_at: string;
-  start_time: number;
-  duration_ms: number | null;
-  server_name: string;
-  upstream_url: string;
-  method: 'POST' | 'GET' | 'DELETE';
-  jsonrpc_method: string | null;
-  tool_name: string | null;
-  api_key: string | null;
-  attribution: string | null;
-  source_ip: string | null;
-  response_status: number | null;
-  is_streamed: boolean;
-  has_debug: boolean;
-  error_code: string | null;
-  error_message: string | null;
-}
-
-export interface McpOAuthSettings {
-  enabled: boolean;
-  provider: 'plexus-idp';
-  issuer?: string;
-}
-
-export interface McpOAuthTokenRecord {
-  id: number;
-  accessTokenHash: string;
-  refreshTokenHash: string;
-  clientId: string;
-  keyName: string;
-  apiKeySecretHash: string | null;
-  resource: string;
-  scope: string | null;
-  accessTokenExpiresAt: number;
-  refreshTokenExpiresAt: number;
-  revokedAt: number | null;
-  createdAt: number;
-}
-
-export interface McpOAuthClientRecord {
-  clientId: string;
-  clientName: string | null;
-  status: 'active' | 'disabled';
-  redirectUris: string[];
-  grantTypes: string[];
-  responseTypes: string[];
-  scope: string | null;
-  tokenEndpointAuthMethod: string;
-  createdAt: number;
-  tokens: McpOAuthTokenRecord[];
-}
-
-export interface LoggingLevelState {
-  level: string;
-  startupLevel: string;
-  supportedLevels: string[];
-  ephemeral: boolean;
-}
-
-export interface ModuleFilterState {
-  modules: string[];
-}
-
-export interface Model {
-  id: string;
-  name: string;
-  providerId: string;
-  pricingSource?: string;
-  type?: 'text' | 'embeddings' | 'transcriptions' | 'speech' | 'image';
-}
-
-// ─── Alias advanced behaviors ────────────────────────────────
-// Mirror of the backend ModelBehaviorSchema discriminated union.
-// Add new variants here as new behavior types are introduced in config.ts.
-
-export interface StripAdaptiveThinkingBehavior {
-  type: 'strip_adaptive_thinking';
-  enabled: boolean;
-}
-
-export type AliasBehavior = StripAdaptiveThinkingBehavior; // | NextBehavior | ...
-
-export type CatalogMetadataSource = 'openrouter' | 'models.dev' | 'catwalk';
-export type MetadataSource = CatalogMetadataSource | 'auto' | 'disabled' | 'custom';
-
-export interface MetadataOverrides {
-  name?: string;
-  description?: string;
-  context_length?: number;
-  pricing?: {
-    prompt?: string;
-    completion?: string;
-    input_cache_read?: string;
-    input_cache_write?: string;
-  };
-  architecture?: {
-    input_modalities?: string[];
-    output_modalities?: string[];
-    tokenizer?: string;
-  };
-  supported_parameters?: string[];
-  top_provider?: {
-    context_length?: number;
-    max_completion_tokens?: number;
-  };
-}
-
-/**
- * Mirror of the backend `NormalizedModelMetadata` shape. Returned by
- * `GET /v1/metadata/lookup` and used to pre-fill the override form.
- */
-export interface NormalizedModelMetadata {
-  id: string;
-  name: string;
-  description?: string;
-  context_length?: number;
-  architecture?: {
-    input_modalities?: string[];
-    output_modalities?: string[];
-    tokenizer?: string;
-    instruct_type?: string | null;
-  };
-  pricing?: {
-    prompt?: string;
-    completion?: string;
-    input_cache_read?: string;
-    input_cache_write?: string;
-  };
-  supported_parameters?: string[];
-  top_provider?: {
-    context_length?: number;
-    max_completion_tokens?: number;
-  };
-}
-
-export interface ModelMetadataRefreshSourceSummary {
-  source: CatalogMetadataSource;
-  initialized: boolean;
-  count: number;
-  error?: string;
-}
-
-export interface ModelMetadataRefreshResult {
-  success: boolean;
-  message: string;
-  trigger: 'startup' | 'scheduled' | 'manual';
-  refreshedAt: string;
-  durationMs: number;
-  intervalMinutes: number;
-  hadErrors: boolean;
-  sources: {
-    openrouter: ModelMetadataRefreshSourceSummary;
-    modelsDev: ModelMetadataRefreshSourceSummary;
-    catwalk: ModelMetadataRefreshSourceSummary;
-  };
-}
-
-// Discriminated union mirrors backend validation: catalog-backed sources
-// must carry a non-empty source_path; 'custom' may omit it but MUST carry
-// an overrides blob with a non-empty `name` (there is no catalog fallback).
-export type AliasMetadata =
-  | {
-      source: CatalogMetadataSource;
-      source_path: string;
-      overrides?: MetadataOverrides;
-    }
-  | {
-      source: 'auto';
-      overrides?: MetadataOverrides;
-    }
-  | {
-      source: 'disabled';
-    }
-  | {
-      source: 'custom';
-      source_path?: string;
-      overrides: MetadataOverrides & { name: string };
-    };
-
-export interface AliasTargetGroup {
-  name: string;
-  selector: string;
-  targets: Array<{
-    provider?: string;
-    model?: string;
-    alias?: string;
-    apiType?: string[];
-    enabled?: boolean;
-  }>;
-}
-
-export type PreferredApiValue = 'chat_completions' | 'messages' | 'gemini' | 'responses';
-
-export interface Alias {
-  id: string;
-  aliases?: string[];
-  priority?: 'selector' | 'api_match';
-  type?: 'text' | 'embeddings' | 'transcriptions' | 'speech' | 'image';
-  target_groups: AliasTargetGroup[];
-  advanced?: AliasBehavior[];
-  metadata?: AliasMetadata;
-  use_image_fallthrough?: boolean;
-  enforce_limits?: boolean;
-  sticky_session?: boolean;
-  preferred_api?: Array<PreferredApiValue>;
-  pi_model?: { provider: string; model_id: string };
-  extraBody?: Record<string, any>;
-  compaction?: CompactionSettings;
-}
-
-export interface InferenceError {
-  id: number;
-  requestId: string;
-  date: string;
-  errorMessage: string;
-  errorStack?: string;
-  details?:
-    | string
-    | {
-        apiType?: string;
-        provider?: string;
-        targetModel?: string;
-        targetApiType?: string;
-        url?: string;
-        headers?: Record<string, string>;
-        statusCode?: number;
-        providerResponse?: string;
-      };
-  createdAt: number;
-}
-
-export interface Cooldown {
-  provider: string;
-  model: string;
-  accountId?: string | null;
-  expiry: number;
-  timeRemainingMs: number;
-  consecutiveFailures?: number;
-  lastError?: string;
-}
-
-// Backend Types
-export interface UsageRecord {
-  requestId: string;
-  clientRequestId?: string | null;
-  date: string;
-  sourceIp?: string;
-  apiKey?: string;
-  attribution?: string;
-  incomingApiType?: string;
-  provider?: string;
-  incomingModelAlias?: string;
-  canonicalModelName?: string | null;
-  selectedModelName?: string;
-  finalAttemptProvider?: string | null;
-  finalAttemptModel?: string | null;
-  allAttemptedProviders?: string | null;
-  outgoingApiType?: string;
-  reasoningEffort?: string | null;
-  tokensInput?: number;
-  tokensOutput?: number;
-  tokensReasoning?: number;
-  tokensCached?: number;
-  tokensCacheWrite?: number;
-  tokensEstimated?: number;
-  costInput?: number;
-  costOutput?: number;
-  costCached?: number;
-  costCacheWrite?: number;
-  costTotal?: number;
-  costSource?: string;
-  costMetadata?: string;
-  startTime: number;
-  durationMs: number | null;
-  isStreamed: boolean;
-  responseStatus: string;
-  ttftMs?: number;
-  tokensPerSec?: number;
-  hasDebug?: boolean;
-  hasError?: boolean;
-  isPassthrough?: boolean;
-  isRaw?: boolean;
-  requestMethod?: string | null;
-  requestPath?: string | null;
-  // Request metadata
-  toolsDefined?: number;
-  messageCount?: number;
-  parallelToolCallsEnabled?: boolean;
-  // Response metadata
-  toolCallsCount?: number;
-  finishReason?: string;
-  // Retry metadata
-  attemptCount?: number;
-  retryHistory?: string | null;
-  // Vision Fallthrough metadata
-  isVisionFallthrough?: boolean;
-  isDescriptorRequest?: boolean;
-  visionFallthroughModel?: string | null;
-  // Provider-reported cost
-  providerReportedCost?: number;
-}
-
-interface BackendResponse<T> {
-  data: T;
-  total: number;
-  error?: string;
-}
-
-interface UsageSummarySeriesPoint {
-  bucketStartMs: number;
-  requests: number;
-  errors: number;
-  inputTokens: number;
-  outputTokens: number;
-  reasoningTokens: number;
-  cachedTokens: number;
-  cacheWriteTokens: number;
-  tokens: number;
-  totalCost: number;
-  avgDurationMs: number;
-  avgTtftMs: number;
-  avgTokensPerSec: number;
-}
-
-export type UsageSummaryBreakdown = 'provider' | 'modelAlias' | 'apiKey' | 'status';
-export type UsageSummaryExclusion = 'directModels' | 'probe';
-
-export interface UsageSummaryGroup {
-  name: string;
-  requests: number;
-  errors: number;
-  inputTokens: number;
-  outputTokens: number;
-  reasoningTokens: number;
-  cachedTokens: number;
-  cacheWriteTokens: number;
-  totalTokens: number;
-  totalCost: number;
-  avgDurationMs: number;
-  totalDurationMs: number;
-  avgTtftMs: number;
-  avgTokensPerSec: number;
-  successRate: number;
-}
-
-export interface UsageSummaryBreakdownResult {
-  items: UsageSummaryGroup[];
-  totalDimensions: number;
-  truncated: boolean;
-}
-
-export interface UsageSummaryResponse {
-  range: 'hour' | 'day' | 'week' | 'month' | 'custom';
-  series: UsageSummarySeriesPoint[];
-  stats: {
-    totalRequests: number;
-    totalErrors: number;
-    totalTokens: number;
-    inputTokens: number;
-    outputTokens: number;
-    reasoningTokens: number;
-    cachedTokens: number;
-    cacheWriteTokens: number;
-    totalCost: number;
-    avgDurationMs: number;
-    totalDurationMs: number;
-    avgTtftMs: number;
-    avgTokensPerSec: number;
-    successRate: number;
-  };
-  today: TodayMetrics;
-  grouped?: Partial<Record<UsageSummaryBreakdown, UsageSummaryBreakdownResult>>;
-}
-
-type UsageRecordField = keyof UsageRecord;
-
-interface UsageQueryParams<T extends UsageRecordField> {
-  limit?: number;
-  offset?: number;
-  requestId?: string;
-  clientRequestId?: string;
-  startDate?: string;
-  endDate?: string;
-  incomingApiType?: string;
-  provider?: string;
-  incomingModelAlias?: string;
-  selectedModelName?: string;
-  outgoingApiType?: string;
-  responseStatus?: string;
-  minDurationMs?: number;
-  maxDurationMs?: number;
-  fields?: T[];
-  cache?: boolean;
-}
-
-const USAGE_CACHE_TTL_MS = 20000;
-const usageRequestCache = new Map<
-  string,
-  { expiresAt: number; promise: Promise<BackendResponse<any>> }
->();
-const summaryRequestCache = new Map<
-  string,
-  { expiresAt: number; promise: Promise<UsageSummaryResponse> }
->();
-
-const CONFIG_CACHE_TTL_MS = 20000;
-const configRequestCache = new Map<string, { expiresAt: number; promise: Promise<any> }>();
-
-export interface QuotaCheckerType {
-  type: string;
-  displayName: string;
-  custom?: boolean;
-}
-
-export interface CustomQuotaChecker {
-  id: string;
-  type: string;
-  displayName: string;
-  code: string;
-  enabled: boolean;
-  createdAt: string;
-  updatedAt: string;
-}
-
-export async function fetchCustomQuotaCheckers(): Promise<CustomQuotaChecker[]> {
-  const response = await fetchWithAuth(`${API_BASE}/v0/management/custom-checkers`);
-  if (!response.ok) throw new Error('Failed to fetch custom quota checkers');
-  return response.json();
-}
-
-export async function saveCustomQuotaChecker(
-  id: string,
-  payload: { displayName: string; code: string; enabled: boolean }
-): Promise<CustomQuotaChecker> {
-  const response = await fetchWithAuth(
-    `${API_BASE}/v0/management/custom-checkers/${encodeURIComponent(id)}`,
-    {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ id, ...payload }),
-    }
-  );
-  if (!response.ok) {
-    const error = await response.json().catch(() => ({}));
-    throw new Error(error.error || 'Failed to save custom quota checker');
-  }
-  return response.json();
-}
-
-export async function deleteCustomQuotaChecker(id: string): Promise<void> {
-  const response = await fetchWithAuth(
-    `${API_BASE}/v0/management/custom-checkers/${encodeURIComponent(id)}`,
-    {
-      method: 'DELETE',
-    }
-  );
-  if (!response.ok) throw new Error('Failed to delete custom quota checker');
-}
-
-export async function testCustomQuotaChecker(
-  id: string,
-  provider: string,
-  options: Record<string, unknown>,
-  code?: string
-): Promise<{ success: boolean; meters: import('../types/quota').Meter[]; error?: string }> {
-  const response = await fetchWithAuth(
-    `${API_BASE}/v0/management/custom-checkers/${encodeURIComponent(id)}/test`,
-    {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ provider, options, ...(code ? { code } : {}) }),
-    }
-  );
-  const result = await response.json();
-  if (!response.ok) throw new Error(result.error || 'Custom quota checker test failed');
-  return result;
-}
-
-export interface QuotaCheckersResponse {
-  knownTypes: QuotaCheckerType[];
-  configured: (QuotaCheckerInfo & { displayName: string; pending: boolean })[];
-}
-
-export async function fetchQuotaCheckers(): Promise<QuotaCheckersResponse> {
-  const response = await fetchWithAuth(`${API_BASE}/v0/management/quota-checkers`);
-  if (!response.ok) throw new Error('Failed to fetch quota checkers');
-  const data = await response.json();
-  return {
-    knownTypes: data.knownTypes ?? [],
-    configured: (data.configured ?? []).map(
-      (c: QuotaCheckerInfo & { displayName: string; pending: boolean }) => ({
-        ...normalizeQuotaCheckerInfo(c),
-        displayName: c.displayName,
-        pending: c.pending,
-      })
-    ),
-  };
-}
-
-const normalizeProviderQuotaChecker = (checker?: {
-  type?: string;
-  enabled?: boolean;
-  intervalMinutes?: number;
-  options?: Record<string, unknown>;
-}): Provider['quotaChecker'] | undefined => {
-  if (!checker) return undefined;
-
-  const type = checker.type?.trim();
-  if (!type) return undefined;
-
-  return {
-    type,
-    enabled: checker.enabled !== false,
-    intervalMinutes: Math.max(1, Number(checker.intervalMinutes || 30)),
-    options: checker.options,
-  };
-};
-
-const USAGE_PAGE_FIELDS: UsageRecordField[] = [
-  'date',
-  'tokensInput',
-  'tokensOutput',
-  'tokensCached',
-  'tokensCacheWrite',
-  'incomingModelAlias',
-  'provider',
-  'apiKey',
-];
-
-const normalizeNow = (): Date => {
-  const now = new Date();
-  now.setSeconds(0, 0);
-  return now;
-};
-
-const getUsageRangeConfig = (range: 'hour' | 'day' | 'week' | 'month' | 'custom', now: Date) => {
-  const startDate = new Date(now);
-  let bucketFormat: (d: Date) => string;
-  let buckets = 0;
-  let step = 0;
-
-  switch (range) {
-    case 'hour':
-      startDate.setHours(startDate.getHours() - 1);
-      bucketFormat = (d) => d.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
-      buckets = 60;
-      step = 60 * 1000;
-      break;
-    case 'day':
-      startDate.setHours(startDate.getHours() - 24);
-      bucketFormat = (d) => d.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
-      buckets = 24;
-      step = 60 * 60 * 1000;
-      break;
-    case 'month':
-      startDate.setDate(startDate.getDate() - 30);
-      bucketFormat = (d) => d.toLocaleDateString();
-      buckets = 30;
-      step = 24 * 60 * 60 * 1000;
-      break;
-    case 'week':
-    default:
-      startDate.setDate(startDate.getDate() - 7);
-      bucketFormat = (d) => d.toLocaleDateString();
-      buckets = 7;
-      step = 24 * 60 * 60 * 1000;
-      break;
-  }
-
-  return { startDate, bucketFormat, buckets, step };
-};
-
-const buildUsageSeries = (
-  records: Array<Partial<UsageRecord>>,
-  range: 'hour' | 'day' | 'week' | 'month',
-  now: Date
-): UsageData[] => {
-  const { startDate, bucketFormat, buckets, step } = getUsageRangeConfig(range, now);
-  const grouped: Record<string, UsageData> = {};
-  const nowMs = now.getTime();
-
-  for (let i = buckets; i >= 0; i--) {
-    const t = new Date(nowMs - i * step);
-    if (range === 'day') t.setMinutes(0, 0, 0);
-    if (range === 'week' || range === 'month') t.setHours(0, 0, 0, 0);
-
-    const key = bucketFormat(t);
-    if (!grouped[key]) {
-      grouped[key] = {
-        timestamp: key,
-        requests: 0,
-        tokens: 0,
-        inputTokens: 0,
-        outputTokens: 0,
-        cachedTokens: 0,
-        cacheWriteTokens: 0,
-      };
-    }
-  }
-
-  records.forEach((record) => {
-    if (!record.date) return;
-    const d = new Date(record.date);
-    if (d < startDate) return;
-
-    if (range === 'day') d.setMinutes(0, 0, 0);
-    if (range === 'week' || range === 'month') d.setHours(0, 0, 0, 0);
-
-    const key = bucketFormat(d);
-    if (!grouped[key]) {
-      grouped[key] = {
-        timestamp: key,
-        requests: 0,
-        tokens: 0,
-        inputTokens: 0,
-        outputTokens: 0,
-        cachedTokens: 0,
-        cacheWriteTokens: 0,
-      };
-    }
-
-    const inputTokens = record.tokensInput || 0;
-    const outputTokens = record.tokensOutput || 0;
-    const cachedTokens = record.tokensCached || 0;
-    const cacheWriteTokens = record.tokensCacheWrite || 0;
-
-    grouped[key].requests++;
-    grouped[key].tokens += inputTokens + outputTokens + cachedTokens + cacheWriteTokens;
-    grouped[key].inputTokens += inputTokens;
-    grouped[key].outputTokens += outputTokens;
-    grouped[key].cachedTokens += cachedTokens;
-    grouped[key].cacheWriteTokens += cacheWriteTokens;
-  });
-
-  return Object.values(grouped);
-};
-
-const formatBucketLabel = (range: 'hour' | 'day' | 'week' | 'month' | 'custom', date: Date) => {
-  if (range === 'hour' || range === 'day') {
-    return date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
-  }
-  return date.toLocaleDateString();
-};
-
-const buildSummarySeries = (summary: UsageSummaryResponse, now: Date): UsageData[] => {
-  const { buckets, step } = getUsageRangeConfig(summary.range, now);
-  const grouped: Record<string, UsageData> = {};
-  const stepMs = step;
-  const alignedNowMs = Math.floor(now.getTime() / stepMs) * stepMs;
-  const startMs = alignedNowMs - buckets * stepMs;
-  const byBucket = new Map(summary.series.map((point) => [point.bucketStartMs, point]));
-
-  for (let i = 0; i <= buckets; i++) {
-    const bucketStartMs = startMs + i * stepMs;
-    const bucketDate = new Date(bucketStartMs);
-    const label = formatBucketLabel(summary.range, bucketDate);
-    const point = byBucket.get(bucketStartMs);
-    const inputTokens = point?.inputTokens || 0;
-    const outputTokens = point?.outputTokens || 0;
-    const cachedTokens = point?.cachedTokens || 0;
-    const cacheWriteTokens = point?.cacheWriteTokens || 0;
-
-    grouped[label] = {
-      timestamp: label,
-      requests: point?.requests || 0,
-      tokens: point?.tokens || inputTokens + outputTokens + cachedTokens + cacheWriteTokens,
-      inputTokens,
-      outputTokens,
-      cachedTokens,
-      cacheWriteTokens,
-    };
-  }
-
-  return Object.values(grouped);
-};
-
-const buildUsageQuery = <T extends UsageRecordField>(params: UsageQueryParams<T>) => {
-  const searchParams = new URLSearchParams();
-
-  if (params.limit !== undefined) searchParams.set('limit', String(params.limit));
-  if (params.offset !== undefined) searchParams.set('offset', String(params.offset));
-  if (params.requestId) searchParams.set('requestId', params.requestId);
-  if (params.clientRequestId) searchParams.set('clientRequestId', params.clientRequestId);
-  if (params.startDate) searchParams.set('startDate', params.startDate);
-  if (params.endDate) searchParams.set('endDate', params.endDate);
-  if (params.incomingApiType) searchParams.set('incomingApiType', params.incomingApiType);
-  if (params.provider) searchParams.set('provider', params.provider);
-  if (params.incomingModelAlias) searchParams.set('incomingModelAlias', params.incomingModelAlias);
-  if (params.selectedModelName) searchParams.set('selectedModelName', params.selectedModelName);
-  if (params.outgoingApiType) searchParams.set('outgoingApiType', params.outgoingApiType);
-  if (params.responseStatus) searchParams.set('responseStatus', params.responseStatus);
-  if (params.minDurationMs !== undefined)
-    searchParams.set('minDurationMs', String(params.minDurationMs));
-  if (params.maxDurationMs !== undefined)
-    searchParams.set('maxDurationMs', String(params.maxDurationMs));
-
-  if (params.fields && params.fields.length > 0) {
-    const fieldsValue = [...params.fields].sort().join(',');
-    searchParams.set('fields', fieldsValue);
-  }
-
-  return searchParams;
-};
-
-const fetchUsageRecords = async <T extends UsageRecordField>(
-  params: UsageQueryParams<T>
-): Promise<BackendResponse<Pick<UsageRecord, T>[]>> => {
-  const searchParams = buildUsageQuery(params);
-  const queryString = searchParams.toString();
-  const url = queryString
-    ? `${API_BASE}/v0/management/usage?${queryString}`
-    : `${API_BASE}/v0/management/usage`;
-
-  if (params.cache) {
-    const cached = usageRequestCache.get(queryString);
-    if (cached && cached.expiresAt > Date.now()) {
-      return cached.promise as Promise<BackendResponse<Pick<UsageRecord, T>[]>>;
-    }
-
-    const promise = (async () => {
-      const res = await fetchWithAuth(url);
-      if (!res.ok) throw new Error('Failed to fetch usage');
-      return (await res.json()) as BackendResponse<Pick<UsageRecord, T>[]>;
-    })();
-
-    usageRequestCache.set(queryString, { expiresAt: Date.now() + USAGE_CACHE_TTL_MS, promise });
-    promise.catch(() => usageRequestCache.delete(queryString));
-    return promise;
-  }
-
-  const res = await fetchWithAuth(url);
-  if (!res.ok) throw new Error('Failed to fetch usage');
-  return (await res.json()) as BackendResponse<Pick<UsageRecord, T>[]>;
-};
-
-const fetchUsageSummary = async (
-  range: 'hour' | 'day' | 'week' | 'month' | 'custom',
-  cache = true,
-  startDate?: string,
-  endDate?: string,
-  breakdowns: UsageSummaryBreakdown[] = [],
-  breakdownLimit = 10,
-  exclusions: UsageSummaryExclusion[] = []
-) => {
-  const searchParams = new URLSearchParams();
-  searchParams.set('range', range);
-
-  if (range === 'custom' && startDate && endDate) {
-    searchParams.set('startDate', startDate);
-    searchParams.set('endDate', endDate);
-  }
-
-  const normalizedBreakdowns = Array.from(new Set(breakdowns)).sort();
-  if (normalizedBreakdowns.length > 0) {
-    searchParams.set('breakdowns', normalizedBreakdowns.join(','));
-    searchParams.set('breakdownLimit', String(breakdownLimit));
-  }
-  const normalizedExclusions = Array.from(new Set(exclusions)).sort();
-  if (normalizedExclusions.length > 0) {
-    searchParams.set('exclude', normalizedExclusions.join(','));
-  }
-
-  const queryString = searchParams.toString();
-  const url = `${API_BASE}/v0/management/usage/summary?${queryString}`;
-
-  if (cache) {
-    const cached = summaryRequestCache.get(queryString);
-    if (cached && cached.expiresAt > Date.now()) {
-      return cached.promise;
-    }
-
-    const promise = (async () => {
-      const res = await fetchWithAuth(url);
-      if (!res.ok) throw new Error('Failed to fetch usage summary');
-      return (await res.json()) as UsageSummaryResponse;
-    })();
-
-    summaryRequestCache.set(queryString, { expiresAt: Date.now() + USAGE_CACHE_TTL_MS, promise });
-    promise.catch(() => summaryRequestCache.delete(queryString));
-    return promise;
-  }
-
-  const res = await fetchWithAuth(url);
-  if (!res.ok) throw new Error('Failed to fetch usage summary');
-  return (await res.json()) as UsageSummaryResponse;
-};
-
-const fetchConfigCached = async (): Promise<any> => {
-  const cached = configRequestCache.get('config');
-  if (cached && cached.expiresAt > Date.now()) {
-    return cached.promise;
-  }
-
-  const promise = (async () => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config`);
-    if (!res.ok) throw new Error('Failed to fetch config');
-    return await res.json();
-  })();
-
-  configRequestCache.set('config', { expiresAt: Date.now() + CONFIG_CACHE_TTL_MS, promise });
-  promise.catch(() => configRequestCache.delete('config'));
-  return promise;
-};
-
-const encodePathPreservingSlashes = (value: string): string =>
-  value.split('/').map(encodeURIComponent).join('/');
-
-export interface KeyConfig {
-  key: string; // The user-facing alias/name for the key (e.g. 'my-app')
-  secret: string; // The actual sk-uuid
-  comment?: string;
-  // Zero or more quota-definition names assigned to this key (non-stacking:
-  // when empty, the system's `default_quotas` fallback applies instead).
-  // The deprecated single `quota` field is still accepted by the backend on
-  // write, but responses only ever emit `quotas` — this type follows suit.
-  quotas?: string[];
-  allowedModels?: string[];
-  allowedProviders?: string[];
-  excludedModels?: string[];
-  excludedProviders?: string[];
-  allowRawPassthrough?: boolean;
-  allowedIps?: string[];
-  expiresInMinutes?: number;
-  expiresAt?: number;
-  disabledAt?: number;
-}
-
-export type UsageSortField =
-  | 'date'
-  | 'apiKey'
-  | 'provider'
-  | 'incomingModelAlias'
-  | 'costTotal'
-  | 'durationMs';
-
-export type UsageSortDirection = 'asc' | 'desc';
-
-/**
- * Scope-restriction fields shared by every quota definition type. Empty (all
- * fields absent/empty) means the quota applies to every provider/model pair.
- * Mirrors `QuotaScopeFields` in `packages/backend/src/config.ts`.
- */
-export interface QuotaScope {
-  allowedProviders?: string[];
-  excludedProviders?: string[];
-  allowedModels?: string[];
-  excludedModels?: string[];
-}
-
-export interface UserQuota extends QuotaScope {
-  type: 'rolling' | 'daily' | 'weekly' | 'monthly';
-  limitType: 'requests' | 'tokens' | 'cost';
-  limit: number;
-  duration?: string; // Required for rolling type
-  // Pools usage across every key referencing this quota into a single bucket
-  // instead of counting each key independently.
-  shared?: boolean;
-  // Early-warning threshold as a fraction of `limit`, e.g. 0.8 = 80%. Must be
-  // in (0, 1) exclusive when set (validated backend-side).
-  warnAt?: number;
-}
-
-/**
- * Wire shape of one entry in the `quotas[]` array returned by
- * `GET /v0/management/quota/status/:key` and `GET /v0/management/self/quota`.
- * Mirrors `QuotaSnapshotJson` in
- * `packages/backend/src/routes/management/_quota-response.ts`.
- */
-export interface QuotaStatusEntry {
-  name: string;
-  limitType: 'requests' | 'tokens' | 'cost';
-  limit: number;
-  currentUsage: number;
-  remaining: number;
-  allowed: boolean;
-  resetsAt: string;
-  scope: QuotaScope;
-  global: boolean;
-  shared: boolean;
-  warnAt?: number;
-  source: 'assigned' | 'default';
-}
-
-export interface QuotaConfig {
-  id: string;
-  type:
-    | 'synthetic'
-    | 'naga'
-    | 'nanogpt'
-    | 'codex'
-    | 'claude-code'
-    | 'zai'
-    | 'moonshot'
-    | 'minimax'
-    | 'minimax-coding'
-    | 'kimi-code'
-    | 'openrouter'
-    | 'kilo';
-  provider: string;
-  enabled: boolean;
-  intervalMinutes: number;
-  options: {
-    apiKey?: string;
-    endpoint?: string;
-    max?: number;
-    oauthProvider?: string;
-    oauthAccountId?: string;
-  };
-  implicit?: boolean;
-}
-
-export interface OAuthProviderInfo {
-  id: string;
-  name: string;
-  usesCallbackServer: boolean;
-}
-
-export interface OAuthAuthInfo {
-  url: string;
-  instructions?: string;
-}
-
-export interface OAuthPrompt {
-  message: string;
-  placeholder?: string;
-  allowEmpty?: boolean;
-}
-
-export interface OAuthSession {
-  id: string;
-  providerId: string;
-  accountId: string;
-  status: string;
-  authInfo?: OAuthAuthInfo;
-  prompt?: OAuthPrompt;
-  progress: string[];
-  error?: string;
-  createdAt: number;
-  updatedAt: number;
-}
-
-export interface OAuthCredentialStatus {
-  ready: boolean;
-}
-
-export const formatLargeNumber = formatNumber;
-export { formatPoints };
-
-export const STAT_LABELS = {
-  REQUESTS: 'Total Requests',
-  PROVIDERS: 'Active Providers',
-  TOKENS: 'Total Tokens',
-  DURATION: 'Avg. Duration',
-} as const;
-
-export interface CompactionSettings {
-  enabled?: boolean;
-  strategy?: 'native' | 'headroom';
-  triggerRatio?: number;
-  absoluteTriggerTokens?: number | null;
-  minTokens?: number;
-  protectRecent?: number;
-  native?: { maxArrayItems?: number; maxStringChars?: number };
-  headroom?: { baseUrl?: string; apiKey?: string; targetRatio?: number | null; timeoutMs?: number };
-}
-
-export interface ModelResolutionPreview {
-  canonical_model: {
-    provider?: string;
-    model: string;
-    basis: 'pi_model' | 'target' | 'alias';
-  };
-  pi_model: { provider: string; model_id: string; name: string } | null;
-  metadata: {
-    source: CatalogMetadataSource | 'heuristic';
-    source_path?: string;
-    name: string;
-  } | null;
-  preferred_api: PreferredApiValue[] | null;
-}
-
-function aliasToConfigPayload(alias: Alias): Record<string, unknown> {
-  const targetGroups = dedupeAliasTargets(alias.target_groups);
-
-  return {
-    priority: alias.priority || 'selector',
-    additional_aliases: dedupeStrings(alias.aliases ?? []),
-    use_image_fallthrough: alias.use_image_fallthrough || false,
-    enforce_limits: alias.enforce_limits || false,
-    sticky_session: alias.sticky_session ?? true,
-    ...(alias.preferred_api?.length ? { preferred_api: alias.preferred_api } : {}),
-    ...(alias.type && { type: alias.type }),
-    ...(alias.advanced?.length ? { advanced: alias.advanced } : {}),
-    ...(alias.metadata && { metadata: alias.metadata }),
-    ...(alias.pi_model && { pi_model: alias.pi_model }),
-    ...(alias.extraBody && Object.keys(alias.extraBody).length > 0
-      ? { extraBody: alias.extraBody }
-      : {}),
-    target_groups: targetGroups.map((group) => ({
-      name: group.name,
-      selector: group.selector,
-      targets: group.targets.map((target) =>
-        target.alias
-          ? {
-              alias: target.alias,
-              ...(target.enabled === false && { enabled: false }),
-            }
-          : {
-              provider: target.provider,
-              model: target.model,
-              ...(target.enabled === false && { enabled: false }),
-            }
-      ),
-    })),
-  };
-}
-
-/**
- * A model entry returned by `GET /v0/management/oauth/models`. Codex is the one
- * OAuth provider whose list is account-scoped and fetched live, so entries can
- * carry a modality hint, the protocols they are reachable through, and the
- * upstream's own listing hint (`hide` models work but are not advertised).
- */
-export interface OAuthDiscoveredModel {
-  id: string;
-  name?: string;
-  context_length?: number;
-  description?: string;
-  pricing?: { prompt?: string; completion?: string };
-  type?: 'text' | 'image';
-  access_via?: string[];
-  visibility?: 'list' | 'hide';
-}
-
-export interface OAuthProviderModelsResult {
-  models: OAuthDiscoveredModel[];
-  /** `catalog` means the live Codex lookup was unavailable and we fell back. */
-  source: 'codex-backend' | 'catalog';
-  /** Present only on a fallback — a warning, not a hard error. */
-  warning?: string;
-}
-
-export const api = {
-  getCooldowns: async (): Promise<Cooldown[]> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/cooldowns`);
-      if (!res.ok) throw new Error('Failed to fetch cooldowns');
-      return await res.json();
-    } catch (e) {
-      console.error('API Error getCooldowns', e);
-      return [];
-    }
-  },
-
-  clearCooldown: async (provider?: string, model?: string): Promise<void> => {
-    let url: string;
-    if (provider) {
-      url = `${API_BASE}/v0/management/cooldowns/${provider}`;
-      if (model) {
-        url += `?model=${encodeURIComponent(model)}`;
-      }
-    } else {
-      url = `${API_BASE}/v0/management/cooldowns`;
-    }
-
-    const res = await fetchWithAuth(url, { method: 'DELETE' });
-    if (!res.ok) throw new Error('Failed to clear cooldown');
-  },
-
-  getStats: async (): Promise<Stat[]> => {
-    try {
-      const [summary, config] = await Promise.all([
-        fetchUsageSummary('week', true),
-        fetchConfigCached(),
-      ]);
-      const activeProviders = config ? Object.keys(config.providers || {}).length : '-';
-
-      const totalRequests = summary.stats.totalRequests || 0;
-      const totalTokens = summary.stats.totalTokens || 0;
-      const avgLatency = Math.round(summary.stats.avgDurationMs || 0);
-
-      return [
-        { label: STAT_LABELS.REQUESTS, value: formatNumber(totalRequests, 0) },
-        { label: STAT_LABELS.PROVIDERS, value: activeProviders },
-        { label: STAT_LABELS.TOKENS, value: formatLargeNumber(totalTokens) },
-        { label: STAT_LABELS.DURATION, value: avgLatency + 'ms' },
-      ];
-    } catch (e) {
-      console.error('API Error getStats', e);
-      return [
-        { label: STAT_LABELS.REQUESTS, value: '-' },
-        { label: STAT_LABELS.PROVIDERS, value: '-' },
-        { label: STAT_LABELS.TOKENS, value: '-' },
-        { label: STAT_LABELS.DURATION, value: '-' },
-      ];
-    }
-  },
-
-  getDashboardData: async (
-    range: 'hour' | 'day' | 'week' | 'month' | 'custom' = 'day',
-    cache = true,
-    startDate?: string,
-    endDate?: string
-  ): Promise<DashboardData> => {
-    try {
-      const now = normalizeNow();
-      const [summary, cooldowns, config] = await Promise.all([
-        fetchUsageSummary(range, cache, startDate, endDate),
-        api.getCooldowns(),
-        fetchConfigCached(),
-      ]);
-
-      const usageData = buildSummarySeries(summary, now);
-      const totalRequests = summary.stats.totalRequests || 0;
-      const totalTokens = summary.stats.totalTokens || 0;
-      const avgLatency = Math.round(summary.stats.avgDurationMs || 0);
-      const activeProviders = config ? Object.keys(config.providers || {}).length : '-';
-
-      const stats: Stat[] = [
-        { label: STAT_LABELS.REQUESTS, value: formatNumber(totalRequests, 0) },
-        { label: STAT_LABELS.PROVIDERS, value: activeProviders },
-        { label: STAT_LABELS.TOKENS, value: formatLargeNumber(totalTokens) },
-        { label: STAT_LABELS.DURATION, value: avgLatency + 'ms' },
-      ];
-
-      return {
-        stats,
-        usageData,
-        cooldowns,
-        todayMetrics: summary.today,
-      };
-    } catch (e) {
-      console.error('API Error getDashboardData', e);
-      return {
-        stats: [
-          { label: STAT_LABELS.REQUESTS, value: '-' },
-          { label: STAT_LABELS.PROVIDERS, value: '-' },
-          { label: STAT_LABELS.TOKENS, value: '-' },
-          { label: STAT_LABELS.DURATION, value: '-' },
-        ],
-        usageData: [],
-        cooldowns: [],
-        todayMetrics: {
-          requests: 0,
-          inputTokens: 0,
-          outputTokens: 0,
-          reasoningTokens: 0,
-          cachedTokens: 0,
-          cacheWriteTokens: 0,
-          totalCost: 0,
-        },
-      };
-    }
-  },
-
-  /**
-   * Fetch the raw usage-summary response for the current principal.
-   *
-   * Unlike `getSummaryData` (which reshapes series into chart-ready rows),
-   * this returns the untransformed backend payload including range-scoped
-   * stats, optional grouped breakdowns, and today roll-ups.
-   *
-   * The backend auto-scopes this endpoint to the calling limited user's
-   * key, so admin callers see global totals and api-key callers see only
-   * their own.
-   */
-  getUsageSummary: async (
-    range: 'hour' | 'day' | 'week' | 'month' | 'custom' = 'day',
-    cache = true,
-    startDate?: string,
-    endDate?: string,
-    breakdowns: UsageSummaryBreakdown[] = [],
-    breakdownLimit = 10,
-    exclusions: UsageSummaryExclusion[] = []
-  ): Promise<UsageSummaryResponse | null> => {
-    try {
-      return await fetchUsageSummary(
-        range,
-        cache,
-        startDate,
-        endDate,
-        breakdowns,
-        breakdownLimit,
-        exclusions
-      );
-    } catch (e) {
-      console.error('API Error getUsageSummary', e);
-      return null;
-    }
-  },
-
-  /**
-   * Fetch pre-aggregated summary data from the backend.
-   * This is much more efficient than getUsageData for time-series views.
-   */
-  getSummaryData: async (
-    range: 'hour' | 'day' | 'week' | 'month' | 'custom' = 'week',
-    cache = true,
-    startDate?: string,
-    endDate?: string,
-    breakdowns: UsageSummaryBreakdown[] = [],
-    breakdownLimit = 10,
-    exclusions: UsageSummaryExclusion[] = []
-  ): Promise<any[]> => {
-    try {
-      const summaryResponse = await fetchUsageSummary(
-        range,
-        cache,
-        startDate,
-        endDate,
-        breakdowns,
-        breakdownLimit,
-        exclusions
-      );
-      const series = summaryResponse.series || [];
-
-      // Return raw data with minimal transformation
-      return series.map((point) => ({
-        timestamp: String(point.bucketStartMs),
-        requests: point.requests,
-        errors: point.errors,
-        tokens: point.tokens,
-        inputTokens: point.inputTokens,
-        outputTokens: point.outputTokens,
-        reasoningTokens: point.reasoningTokens,
-        cachedTokens: point.cachedTokens,
-        cacheWriteTokens: point.cacheWriteTokens,
-        cost: point.totalCost,
-        duration: point.avgDurationMs,
-        ttft: point.avgTtftMs,
-        tps: point.avgTokensPerSec,
-      }));
-    } catch (e) {
-      console.error('API Error getSummaryData', e);
-      return [];
-    }
-  },
-
-  getUsageData: async (
-    range: 'hour' | 'day' | 'week' | 'month' | 'custom' = 'week',
-    cache = true,
-    startDate?: string,
-    endDate?: string
-  ): Promise<UsageData[]> => {
-    try {
-      const now = normalizeNow();
-
-      let queryStartDate: Date;
-      let queryEndDate: Date;
-
-      if (range === 'custom' && startDate && endDate) {
-        queryStartDate = new Date(startDate);
-        queryEndDate = new Date(endDate);
-      } else {
-        const { startDate: configStart } = getUsageRangeConfig(
-          range as 'hour' | 'day' | 'week' | 'month',
-          now
-        );
-        queryStartDate = configStart;
-        queryEndDate = now;
-      }
-
-      const usageResponse = await fetchUsageRecords({
-        limit: 5000,
-        startDate: queryStartDate.toISOString(),
-        endDate: queryEndDate.toISOString(),
-        fields: USAGE_PAGE_FIELDS,
-        cache,
-      });
-
-      return buildUsageSeries(usageResponse.data || [], range === 'custom' ? 'day' : range, now);
-    } catch (e) {
-      console.error('API Error getUsageData', e);
-      return [];
-    }
-  },
-
-  getTodayMetrics: async (): Promise<TodayMetrics> => {
-    try {
-      const now = normalizeNow();
-      const startDate = new Date(now);
-      startDate.setHours(0, 0, 0, 0);
-
-      const usageResponse = await fetchUsageRecords({
-        limit: 5000,
-        startDate: startDate.toISOString(),
-        fields: [
-          'date',
-          'tokensInput',
-          'tokensOutput',
-          'tokensReasoning',
-          'tokensCached',
-          'tokensCacheWrite',
-          'costTotal',
-        ],
-        cache: true,
-      });
-
-      const metrics: TodayMetrics = {
-        requests: 0,
-        inputTokens: 0,
-        outputTokens: 0,
-        reasoningTokens: 0,
-        cachedTokens: 0,
-        cacheWriteTokens: 0,
-        totalCost: 0,
-      };
-
-      (usageResponse.data || []).forEach((r) => {
-        metrics.requests++;
-        metrics.inputTokens += r.tokensInput || 0;
-        metrics.outputTokens += r.tokensOutput || 0;
-        metrics.reasoningTokens += r.tokensReasoning || 0;
-        metrics.cachedTokens += r.tokensCached || 0;
-        metrics.cacheWriteTokens += r.tokensCacheWrite || 0;
-        metrics.totalCost += r.costTotal || 0;
-      });
-
-      return metrics;
-    } catch (e) {
-      console.error('API Error getTodayMetrics', e);
-      return {
-        requests: 0,
-        inputTokens: 0,
-        outputTokens: 0,
-        reasoningTokens: 0,
-        cachedTokens: 0,
-        cacheWriteTokens: 0,
-        totalCost: 0,
-      };
-    }
-  },
-
-  getProviderPerformance: async (
-    model?: string,
-    provider?: string
-  ): Promise<ProviderPerformanceData[]> => {
-    try {
-      const params = new URLSearchParams();
-      if (model) params.set('model', model);
-      if (provider) params.set('provider', provider);
-
-      const query = params.toString();
-      const url = `${API_BASE}/v0/management/performance${query ? `?${query}` : ''}`;
-
-      const res = await fetchWithAuth(url);
-      if (!res.ok) throw new Error('Failed to fetch provider performance');
-
-      const rawRows = (await res.json()) as Array<Record<string, unknown>>;
-
-      const toNumber = (value: unknown): number => {
-        const parsed = Number(value);
-        return Number.isFinite(parsed) ? parsed : 0;
-      };
-
-      return rawRows.map((row) => ({
-        provider: String(row.provider ?? ''),
-        model: String(row.model ?? ''),
-        target_model: row.target_model ? String(row.target_model) : undefined,
-        avg_ttft_ms: toNumber(row.avg_ttft_ms),
-        min_ttft_ms: toNumber(row.min_ttft_ms),
-        max_ttft_ms: toNumber(row.max_ttft_ms),
-        avg_tokens_per_sec: toNumber(row.avg_tokens_per_sec),
-        min_tokens_per_sec: toNumber(row.min_tokens_per_sec),
-        max_tokens_per_sec: toNumber(row.max_tokens_per_sec),
-        avg_e2e_tokens_per_sec: toNumber(row.avg_e2e_tokens_per_sec),
-        min_e2e_tokens_per_sec: toNumber(row.min_e2e_tokens_per_sec),
-        max_e2e_tokens_per_sec: toNumber(row.max_e2e_tokens_per_sec),
-        sample_count: toNumber(row.sample_count),
-        last_updated: toNumber(row.last_updated),
-      }));
-    } catch (e) {
-      console.error('API Error getProviderPerformance', e);
-      return [];
-    }
-  },
-
-  clearProviderPerformance: async (model: string): Promise<boolean> => {
-    try {
-      const url = `${API_BASE}/v0/management/performance?model=${encodeURIComponent(model)}`;
-      const res = await fetchWithAuth(url, { method: 'DELETE' });
-      if (!res.ok) throw new Error('Failed to clear provider performance');
-      return true;
-    } catch (e) {
-      console.error('API Error clearProviderPerformance', e);
-      return false;
-    }
-  },
-
-  getLogs: async (
-    limit: number = 50,
-    offset: number = 0,
-    filters: Record<string, any> = {},
-    sortBy: UsageSortField = 'date',
-    sortDir: UsageSortDirection = 'desc'
-  ): Promise<{ data: UsageRecord[]; total: number }> => {
-    const params = new URLSearchParams({
-      limit: limit.toString(),
-      offset: offset.toString(),
-      sortBy,
-      sortDir,
-      ...filters,
-    });
-
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/usage?${params}`);
-    if (!res.ok) throw new Error('Failed to fetch logs');
-    return (await res.json()) as BackendResponse<UsageRecord[]>;
-  },
-
-  getUsageRecords: fetchUsageRecords,
-
-  getConfig: async (): Promise<any> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config`);
-    if (!res.ok) throw new Error('Failed to fetch config');
-    return await res.json();
-  },
-
-  getConfigExport: async (): Promise<any> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/export`);
-    if (!res.ok) throw new Error('Failed to fetch config export');
-    return await res.json();
-  },
-
-  /** All system settings (a flat key/value bag; `default_quotas` lives here). */
-  // getSystemSettings + patchSystemSettings are defined further below (upstream)
-
-  /** Convenience wrapper: the key names substituted in when a key has no
-   * `quotas` of its own. */
-  getDefaultQuotas: async (): Promise<string[]> => {
-    const settings = await api.getSystemSettings();
-    return Array.isArray(settings.default_quotas)
-      ? settings.default_quotas.filter((v): v is string => typeof v === 'string')
-      : [];
-  },
-
-  setDefaultQuotas: async (names: string[]): Promise<void> => {
-    await api.patchSystemSettings({ default_quotas: names });
-  },
-
-  restart: async (): Promise<{ success: boolean; message: string }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/restart`, {
-      method: 'POST',
-    });
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({ error: 'Failed to restart' }));
-      throw new Error(err.error || 'Failed to restart');
-    }
-    return res.json();
-  },
-
-  getKeys: async (): Promise<KeyConfig[]> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/keys`);
-      if (!res.ok) throw new Error('Failed to fetch keys');
-      const keys = (await res.json()) as Record<
-        string,
-        {
-          secret: string;
-          comment?: string;
-          quotas?: string[];
-          allowedModels?: string[];
-          allowedProviders?: string[];
-          excludedModels?: string[];
-          excludedProviders?: string[];
-          allowRawPassthrough?: boolean;
-          allowedIps?: string[];
-          expiresAt?: number;
-          disabledAt?: number;
-        }
-      >;
-
-      return Object.entries(keys).map(([key, val]) => ({
-        key,
-        secret: val.secret,
-        comment: val.comment,
-        quotas: val.quotas,
-        allowedModels: val.allowedModels,
-        allowedProviders: val.allowedProviders,
-        excludedModels: val.excludedModels,
-        excludedProviders: val.excludedProviders,
-        allowRawPassthrough: val.allowRawPassthrough === true,
-        allowedIps: val.allowedIps,
-        expiresAt: val.expiresAt,
-        disabledAt: val.disabledAt,
-      }));
-    } catch (e) {
-      console.error('API Error getKeys', e);
-      return [];
-    }
-  },
-
-  saveKey: async (keyConfig: KeyConfig, oldKeyName?: string): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/keys/${encodeURIComponent(keyConfig.key)}`,
-      {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-          secret: keyConfig.secret,
-          comment: keyConfig.comment,
-          quotas: keyConfig.quotas ?? [],
-          allowedModels: keyConfig.allowedModels ?? [],
-          allowedProviders: keyConfig.allowedProviders ?? [],
-          excludedModels: keyConfig.excludedModels ?? [],
-          excludedProviders: keyConfig.excludedProviders ?? [],
-          allowRawPassthrough: keyConfig.allowRawPassthrough === true,
-          allowedIps: keyConfig.allowedIps ?? [],
-          ...(keyConfig.expiresInMinutes ? { expiresInMinutes: keyConfig.expiresInMinutes } : {}),
-        }),
-      }
-    );
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      const detail =
-        Array.isArray(err.details) && err.details[0]?.message ? `: ${err.details[0].message}` : '';
-      throw new Error(`${err.error || 'Failed to save key'}${detail}`);
-    }
-
-    // Delete old key only after new one is saved successfully
-    if (oldKeyName && oldKeyName !== keyConfig.key) {
-      await api.deleteKey(oldKeyName);
-    }
-  },
-
-  disableKey: async (keyName: string): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/keys/${encodeURIComponent(keyName)}/disable`,
-      { method: 'POST' }
-    );
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to disable key');
-    }
-  },
-
-  deleteKey: async (keyName: string): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/keys/${encodeURIComponent(keyName)}`,
-      { method: 'DELETE' }
-    );
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to delete key');
-    }
-  },
-
-  getProviders: async (): Promise<Provider[]> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/providers`);
-      if (!res.ok) throw new Error('Failed to fetch providers');
-      const providers = (await res.json()) as Record<string, any>;
-
-      return Object.entries(providers).map(([key, val]) => {
-        // Normalize models array format to object format
-        let normalizedModels = val.models;
-        if (Array.isArray(val.models)) {
-          normalizedModels = val.models.reduce(
-            (acc: Record<string, any>, modelName: string) => {
-              acc[modelName] = {};
-              return acc;
-            },
-            {} as Record<string, any>
-          );
-        }
-
-        // Infer type from api_base_url if not explicitly provided
-        const inferredTypes = val.type || inferProviderTypes(val.api_base_url);
-
-        return {
-          id: key,
-          name: val.display_name || key,
-          type: inferredTypes,
-          apiBaseUrl: val.api_base_url,
-          apiKey: val.api_key || '',
-          oauthProvider: val.oauth_provider,
-          oauthAccount: val.oauth_account,
-          enabled: val.enabled !== false,
-          estimateTokens: val.estimateTokens || false,
-          useClaudeMasking: val.useClaudeMasking === true,
-          geminiThinkingEnabled: val.gemini_thinking_enabled === true,
-          disableCooldown: val.disable_cooldown === true,
-          stallCooldown: val.stall_cooldown === true,
-          allow100PercentUtilization: val.allow_100_percent_utilization === true,
-          auto_compat: val.auto_compat === true,
-          discount: val.discount,
-          headers: val.headers,
-          extraBody:
-            val.extraBody && typeof val.extraBody === 'object' && !Array.isArray(val.extraBody)
-              ? val.extraBody
-              : {},
-          models: normalizedModels,
-          quotaChecker: normalizeProviderQuotaChecker(val.quota_checker),
-          modelAutosync: val.model_autosync
-            ? {
-                enabled: val.model_autosync.enabled === true,
-                intervalMinutes: Math.max(1, val.model_autosync.intervalMinutes || 60),
-              }
-            : { enabled: false, intervalMinutes: 60 },
-          adapter: val.adapter ? (Array.isArray(val.adapter) ? val.adapter : [val.adapter]) : [],
-          timeoutMs: val.timeoutMs ?? undefined,
-          maxConcurrency: val.maxConcurrency ?? undefined,
-          stallTtfbMs: val.stallTtfbMs ?? undefined,
-          stallTtfbBytes: val.stallTtfbBytes ?? undefined,
-          stallMinBps: val.stallMinBps ?? undefined,
-          stallWindowMs: val.stallWindowMs ?? undefined,
-          stallGracePeriodMs: val.stallGracePeriodMs ?? undefined,
-          pi_ai_provider: val.pi_ai_provider ?? undefined,
-          rawPassthrough: val.raw_passthrough
-            ? {
-                enabled: val.raw_passthrough.enabled === true,
-                baseUrl: val.raw_passthrough.base_url || '',
-                auth: val.raw_passthrough.auth || 'bearer',
-              }
-            : undefined,
-        };
-      });
-    } catch (e) {
-      console.error('API Error getProviders', e);
-      return [];
-    }
-  },
-
-  saveProvider: async (provider: Provider, oldId?: string): Promise<void> => {
-    const body: any = {
-      api_base_url: provider.apiBaseUrl,
-      display_name: provider.name,
-      api_key: provider.apiKey,
-      ...(provider.oauthProvider && { oauth_provider: provider.oauthProvider }),
-      ...(provider.oauthAccount && { oauth_account: provider.oauthAccount }),
-      enabled: provider.enabled,
-      estimateTokens: provider.estimateTokens,
-      useClaudeMasking: provider.useClaudeMasking,
-      geminiThinkingEnabled: provider.geminiThinkingEnabled,
-      disable_cooldown: provider.disableCooldown === true,
-      stall_cooldown: provider.stallCooldown === true,
-      allow_100_percent_utilization: provider.allow100PercentUtilization === true,
-      auto_compat: provider.auto_compat === true,
-      discount: provider.discount,
-      headers: provider.headers,
-      extraBody: provider.extraBody,
-      models: provider.models,
-      quota_checker: provider.quotaChecker?.type
-        ? {
-            type: provider.quotaChecker.type,
-            enabled: provider.quotaChecker.enabled,
-            intervalMinutes: Math.max(1, provider.quotaChecker.intervalMinutes || 30),
-            options: provider.quotaChecker.options,
-          }
-        : undefined,
-      model_autosync: {
-        enabled: provider.modelAutosync?.enabled === true,
-        intervalMinutes: Math.max(1, provider.modelAutosync?.intervalMinutes || 60),
-      },
-      adapter: provider.adapter ?? [],
-      ...(provider.timeoutMs != null ? { timeoutMs: provider.timeoutMs } : {}),
-      ...(provider.maxConcurrency != null ? { maxConcurrency: provider.maxConcurrency } : {}),
-      ...(provider.stallTtfbMs != null ? { stallTtfbMs: provider.stallTtfbMs } : {}),
-      ...(provider.stallTtfbBytes != null ? { stallTtfbBytes: provider.stallTtfbBytes } : {}),
-      ...(provider.stallMinBps != null ? { stallMinBps: provider.stallMinBps } : {}),
-      ...(provider.stallWindowMs != null ? { stallWindowMs: provider.stallWindowMs } : {}),
-      ...(provider.stallGracePeriodMs != null
-        ? { stallGracePeriodMs: provider.stallGracePeriodMs }
-        : {}),
-      ...(provider.rawPassthrough?.baseUrl
-        ? {
-            raw_passthrough: {
-              enabled: provider.rawPassthrough.enabled,
-              base_url: provider.rawPassthrough.baseUrl,
-              auth: provider.rawPassthrough.auth,
-            },
-          }
-        : {}),
-      ...(provider.pi_ai_provider ? { pi_ai_provider: provider.pi_ai_provider } : {}),
-    };
-
-    const isExistingProvider = oldId === provider.id;
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/providers/${encodePathPreservingSlashes(provider.id)}`,
-      {
-        method: isExistingProvider ? 'PATCH' : 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(body),
-      }
-    );
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      const detail = err.details ? ` — ${JSON.stringify(err.details)}` : '';
-      throw new Error((err.error || 'Failed to save provider') + detail);
-    }
-
-    // Delete old provider only after new one is saved successfully
-    if (oldId && oldId !== provider.id) {
-      await api.deleteProvider(oldId, false);
-    }
-  },
-
-  updateProviderEnabled: async (providerId: string, enabled: boolean): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/providers/${encodePathPreservingSlashes(providerId)}`,
-      {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ enabled }),
-      }
-    );
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to update provider status');
-    }
-  },
-
-  getVisionFallthroughConfig: async (): Promise<{
-    descriptor_model?: string;
-    default_prompt?: string;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/vision-fallthrough`);
-    if (!res.ok) throw new Error('Failed to fetch vision fallthrough config');
-    return res.json();
-  },
-
-  updateVisionFallthroughConfig: async (updates: {
-    descriptor_model?: string;
-    default_prompt?: string;
-  }): Promise<any> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/vision-fallthrough`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(updates),
-    });
-    if (!res.ok) throw new Error('Failed to update vision fallthrough config');
-    return res.json();
-  },
-
-  deleteProvider: async (
-    providerId: string,
-    cascade?: boolean
-  ): Promise<{
-    success: boolean;
-    provider: string;
-    removedTargets?: number;
-    affectedAliases?: string[];
-  }> => {
-    try {
-      const url = `/v0/management/providers/${encodePathPreservingSlashes(providerId)}${cascade ? '?cascade=true' : ''}`;
-
-      const response = await fetchWithAuth(url, {
-        method: 'DELETE',
-      });
-
-      if (!response.ok) {
-        const error = await response.json().catch(() => ({ error: 'Failed to delete provider' }));
-        throw new Error(error.error || `HTTP ${response.status}`);
-      }
-
-      return await response.json();
-    } catch (e) {
-      console.error('API Error deleteProvider', e);
-      throw e;
-    }
-  },
-
-  getAffectedAliases: async (
-    providerId: string
-  ): Promise<{ aliasId: string; targetsCount: number }[]> => {
-    try {
-      const aliases = await api.getAliases();
-      const affected: { aliasId: string; targetsCount: number }[] = [];
-
-      for (const alias of aliases) {
-        const targetsCount = alias.target_groups.reduce(
-          (sum, g) => sum + g.targets.filter((t) => t.provider === providerId).length,
-          0
-        );
-        if (targetsCount > 0) {
-          affected.push({ aliasId: alias.id, targetsCount });
-        }
-      }
-
-      return affected;
-    } catch (e) {
-      console.error('API Error getAffectedAliases', e);
-      return [];
-    }
-  },
-
-  saveAlias: async (alias: Alias, oldId?: string): Promise<void> => {
-    const body = aliasToConfigPayload(alias);
-
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/aliases/${encodeURIComponent(alias.id)}`,
-      {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(body),
-      }
-    );
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      const detail =
-        Array.isArray(err.details) && err.details[0]?.message ? `: ${err.details[0].message}` : '';
-      throw new Error(`${err.error || 'Failed to save alias'}${detail}`);
-    }
-
-    // Delete old alias only after new one is saved successfully
-    if (oldId && oldId !== alias.id) {
-      await api.deleteAlias(oldId);
-    }
-  },
-
-  previewModelResolution: async (alias: Alias): Promise<ModelResolutionPreview> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/models/metadata/resolve`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ alias_id: alias.id, model: aliasToConfigPayload(alias) }),
-    });
-    if (!res.ok) throw new Error('Failed to resolve automatic model selections');
-    return res.json();
-  },
-
-  getModels: async (): Promise<Model[]> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/providers`);
-      if (!res.ok) throw new Error('Failed to fetch providers');
-      const providers = (await res.json()) as Record<string, any>;
-      const models: Model[] = [];
-
-      // Extract models from providers
-      Object.entries(providers).forEach(([pKey, pVal]) => {
-        if (pVal.models) {
-          if (Array.isArray(pVal.models)) {
-            pVal.models.forEach((m: string) => {
-              models.push({
-                id: m,
-                name: m,
-                providerId: pKey,
-              });
-            });
-          } else if (typeof pVal.models === 'object') {
-            Object.entries(pVal.models).forEach(([mKey, mVal]: [string, any]) => {
-              models.push({
-                id: mKey,
-                name: mKey,
-                providerId: pKey,
-                pricingSource: mVal.pricing?.source,
-                type: mVal.type,
-              });
-            });
-          }
-        }
-      });
-      return dedupeModels(models);
-    } catch (e) {
-      console.error('API Error getModels', e);
-      return [];
-    }
-  },
-
-  getAliases: async (): Promise<Alias[]> => {
-    try {
-      const [aliasRes, providerRes] = await Promise.all([
-        fetchWithAuth(`${API_BASE}/v0/management/aliases`),
-        fetchWithAuth(`${API_BASE}/v0/management/providers`),
-      ]);
-      if (!aliasRes.ok) throw new Error('Failed to fetch aliases');
-      if (!providerRes.ok) throw new Error('Failed to fetch providers');
-
-      const aliasMap = (await aliasRes.json()) as Record<string, any>;
-      const providers = (await providerRes.json()) as Record<string, any>;
-      const aliases: Alias[] = [];
-
-      Object.entries(aliasMap).forEach(([key, val]) => {
-        const readTarget = (t: any) => {
-          if (t.alias) {
-            return {
-              alias: t.alias as string,
-              enabled: t.enabled !== false,
-            };
-          }
-
-          const providerConfig = providers[t.provider];
-          const inferredTypes =
-            providerConfig?.type || inferProviderTypes(providerConfig?.api_base_url);
-          let apiType: string | string[] = inferredTypes;
-          if (providerConfig?.models && !Array.isArray(providerConfig.models)) {
-            const modelConfig = providerConfig.models[t.model];
-            if (modelConfig?.access_via?.length > 0) {
-              apiType = normalizeApiAccessList(modelConfig.access_via);
-            }
-          }
-          return {
-            provider: t.provider,
-            model: t.model,
-            apiType: Array.isArray(apiType) ? apiType : [apiType],
-            enabled: t.enabled !== false,
-          };
-        };
-
-        const targetGroups: AliasTargetGroup[] = (val.target_groups || []).map((g: any) => ({
-          name: g.name || 'default',
-          selector: g.selector || 'random',
-          targets: (g.targets || []).map(readTarget),
-        }));
-
-        aliases.push({
-          id: key,
-          aliases: val.additional_aliases || [],
-          priority: val.priority,
-          type: val.type,
-          target_groups: targetGroups,
-          use_image_fallthrough: val.use_image_fallthrough || false,
-          enforce_limits: val.enforce_limits || false,
-          sticky_session: val.sticky_session ?? true,
-          advanced: val.advanced || [],
-          metadata: val.metadata,
-          preferred_api: val.preferred_api || [],
-          pi_model: val.pi_model,
-          extraBody:
-            val.extraBody && typeof val.extraBody === 'object' && !Array.isArray(val.extraBody)
-              ? val.extraBody
-              : {},
-        });
-      });
-      return aliases;
-    } catch (e) {
-      console.error('API Error getAliases', e);
-      return [];
-    }
-  },
-
-  getDebugLogs: async (
-    limit: number = 50,
-    offset: number = 0
-  ): Promise<{ requestId: string; createdAt: number; responseStatus: number | null }[]> => {
-    try {
-      const res = await fetchWithAuth(
-        `${API_BASE}/v0/management/debug/logs?limit=${limit}&offset=${offset}`
-      );
-      if (!res.ok) throw new Error('Failed to fetch debug logs');
-      const json = (await res.json()) as BackendResponse<
-        { requestId: string; createdAt: number; responseStatus: number | null }[]
-      >;
-      return json.data || [];
-    } catch (e) {
-      console.error('API Error getDebugLogs', e);
-      return [];
-    }
-  },
-
-  getDebugLogDetail: async (requestId: string): Promise<any> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/debug/logs/${requestId}`);
-      if (!res.ok) throw new Error('Failed to fetch debug log detail');
-      return await res.json();
-    } catch (e) {
-      console.error('API Error getDebugLogDetail', e);
-      return null;
-    }
-  },
-
-  deleteDebugLog: async (requestId: string): Promise<boolean> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/debug/logs/${requestId}`, {
-        method: 'DELETE',
-      });
-      return res.ok;
-    } catch (e) {
-      console.error('API Error deleteDebugLog', e);
-      return false;
-    }
-  },
-
-  deleteAllDebugLogs: async (): Promise<boolean> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/debug/logs`, {
-        method: 'DELETE',
-      });
-      return res.ok;
-    } catch (e) {
-      console.error('API Error deleteAllDebugLogs', e);
-      return false;
-    }
-  },
-
-  getErrors: async (limit: number = 50, offset: number = 0): Promise<InferenceError[]> => {
-    try {
-      const res = await fetchWithAuth(
-        `${API_BASE}/v0/management/errors?limit=${limit}&offset=${offset}`
-      );
-      if (!res.ok) throw new Error('Failed to fetch error logs');
-      const json = (await res.json()) as BackendResponse<InferenceError[]>;
-      return json.data || [];
-    } catch (e) {
-      console.error('API Error getErrors', e);
-      return [];
-    }
-  },
-
-  deleteError: async (requestId: string): Promise<boolean> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/errors/${requestId}`, {
-        method: 'DELETE',
-      });
-      return res.ok;
-    } catch (e) {
-      console.error('API Error deleteError', e);
-      return false;
-    }
-  },
-
-  deleteAllErrors: async (): Promise<boolean> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/errors`, {
-        method: 'DELETE',
-      });
-      return res.ok;
-    } catch (e) {
-      console.error('API Error deleteAllErrors', e);
-      return false;
-    }
-  },
-
-  deleteUsageLog: async (requestId: string): Promise<boolean> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/usage/${requestId}`, {
-        method: 'DELETE',
-      });
-      return res.ok;
-    } catch (e) {
-      console.error('API Error deleteUsageLog', e);
-      return false;
-    }
-  },
-
-  deleteAllUsageLogs: async (olderThanDays?: number): Promise<boolean> => {
-    try {
-      let url = `${API_BASE}/v0/management/usage`;
-      if (olderThanDays !== undefined) {
-        url += `?olderThanDays=${olderThanDays}`;
-      }
-      const res = await fetchWithAuth(url, {
-        method: 'DELETE',
-      });
-      return res.ok;
-    } catch (e) {
-      console.error('API Error deleteAllUsageLogs', e);
-      return false;
-    }
-  },
-
-  getDebugMode: async (): Promise<{
-    enabled: boolean;
-    enabledGlobal?: boolean;
-    enabledKeys?: string[];
-    providers: string[] | null;
-    keys: string[] | null;
-    aliases: string[] | null;
-  }> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/debug`);
-      if (!res.ok) throw new Error('Failed to fetch debug status');
-      const json = await res.json();
-      return {
-        enabled: !!json.enabled,
-        enabledGlobal: json.enabledGlobal === undefined ? undefined : !!json.enabledGlobal,
-        enabledKeys: Array.isArray(json.enabledKeys) ? json.enabledKeys : undefined,
-        providers: json.providers || null,
-        keys: json.keys || json.enabledKeys || null,
-        aliases: json.aliases || null,
-      };
-    } catch (e) {
-      console.error('API Error getDebugMode', e);
-      return { enabled: false, providers: null, keys: null, aliases: null };
-    }
-  },
-
-  setDebugMode: async (
-    enabled: boolean,
-    providers?: string[] | null,
-    keys?: string[] | null,
-    aliases?: string[] | null
-  ): Promise<{
-    enabled: boolean;
-    enabledGlobal?: boolean;
-    enabledKeys?: string[];
-    providers: string[] | null;
-    keys: string[] | null;
-    aliases: string[] | null;
-  }> => {
-    try {
-      const body: {
-        enabled: boolean;
-        providers?: string[] | null;
-        keys?: string[] | null;
-        aliases?: string[] | null;
-      } = { enabled };
-      if (providers !== undefined) {
-        body.providers = providers;
-      }
-      if (keys !== undefined) {
-        body.keys = keys;
-      }
-      if (aliases !== undefined) {
-        body.aliases = aliases;
-      }
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/debug`, {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(body),
-      });
-      if (!res.ok) throw new Error('Failed to set debug status');
-      const json = await res.json();
-      return {
-        enabled: !!json.enabled,
-        enabledGlobal: json.enabledGlobal === undefined ? undefined : !!json.enabledGlobal,
-        enabledKeys: Array.isArray(json.enabledKeys) ? json.enabledKeys : undefined,
-        providers: json.providers || null,
-        keys: json.keys || json.enabledKeys || null,
-        aliases: json.aliases || null,
-      };
-    } catch (e) {
-      console.error('API Error setDebugMode', e);
-      throw e;
-    }
-  },
-
-  getLoggingLevel: async (): Promise<LoggingLevelState> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/logging/level`);
-      if (!res.ok) throw new Error('Failed to fetch logging level');
-      const json = (await res.json()) as LoggingLevelState;
-      return {
-        level: json.level,
-        startupLevel: json.startupLevel,
-        supportedLevels: Array.isArray(json.supportedLevels)
-          ? json.supportedLevels
-          : ['error', 'warn', 'info', 'debug', 'verbose', 'silly'],
-        ephemeral: !!json.ephemeral,
-      };
-    } catch (e) {
-      console.error('API Error getLoggingLevel', e);
-      return {
-        level: 'info',
-        startupLevel: 'info',
-        supportedLevels: ['error', 'warn', 'info', 'debug', 'verbose', 'silly'],
-        ephemeral: true,
-      };
-    }
-  },
-
-  setLoggingLevel: async (level: string): Promise<LoggingLevelState> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/logging/level`, {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ level }),
-    });
-
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to set logging level');
-    }
-
-    const json = (await res.json()) as LoggingLevelState;
-    return {
-      level: json.level,
-      startupLevel: json.startupLevel,
-      supportedLevels: Array.isArray(json.supportedLevels)
-        ? json.supportedLevels
-        : ['error', 'warn', 'info', 'debug', 'verbose', 'silly'],
-      ephemeral: !!json.ephemeral,
-    };
-  },
-
-  resetLoggingLevel: async (): Promise<LoggingLevelState> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/logging/level`, {
-      method: 'DELETE',
-    });
-
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to reset logging level');
-    }
-
-    const json = (await res.json()) as LoggingLevelState;
-    return {
-      level: json.level,
-      startupLevel: json.startupLevel,
-      supportedLevels: Array.isArray(json.supportedLevels)
-        ? json.supportedLevels
-        : ['error', 'warn', 'info', 'debug', 'verbose', 'silly'],
-      ephemeral: !!json.ephemeral,
-    };
-  },
-
-  getModuleFilter: async (): Promise<ModuleFilterState> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/logging/modules`);
-    if (!res.ok) throw new Error('Failed to fetch module filter');
-    const json = (await res.json()) as ModuleFilterState;
-    return { modules: Array.isArray(json.modules) ? json.modules : [] };
-  },
-
-  setModuleFilter: async (modules: string[]): Promise<ModuleFilterState> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/logging/modules`, {
-      method: 'PUT',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ modules }),
-    });
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to set module filter');
-    }
-    const json = (await res.json()) as ModuleFilterState;
-    return { modules: Array.isArray(json.modules) ? json.modules : [] };
-  },
-
-  clearModuleFilter: async (): Promise<void> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/logging/modules`, {
-      method: 'DELETE',
-    });
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to clear module filter');
-    }
-  },
-
-  testModel: async (
-    provider: string,
-    model: string,
-    apiType?: string
-  ): Promise<{
-    success: boolean;
-    error?: string;
-    durationMs: number | null;
-    response?: string;
-    apiType?: string;
-  }> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/test`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ provider, model, apiType }),
-      });
-      if (!res.ok) throw new Error('Failed to test model');
-      return await res.json();
-    } catch (e) {
-      console.error('API Error testModel', e);
-      throw e;
-    }
-  },
-
-  getQuotas: async (): Promise<QuotaCheckerInfo[]> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/quotas`);
-      if (!res.ok) throw new Error('Failed to fetch quotas');
-      const json = (await res.json()) as QuotaCheckerInfo[];
-      return Array.isArray(json) ? json.map(normalizeQuotaCheckerInfo) : [];
-    } catch (e) {
-      console.error('API Error getQuotas', e);
-      return [];
-    }
-  },
-
-  getQuota: async (checkerId: string): Promise<QuotaCheckerInfo | null> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/quotas/${checkerId}`);
-      if (!res.ok) throw new Error('Failed to fetch quota');
-      const json = (await res.json()) as QuotaCheckerInfo;
-      return normalizeQuotaCheckerInfo(json);
-    } catch (e) {
-      console.error('API Error getQuota', e);
-      return null;
-    }
-  },
-
-  getQuotaHistory: async (
-    checkerId: string,
-    meterKey?: string,
-    since?: string
-  ): Promise<{ checkerId: string; meterKey?: string; since?: string; history: any[] } | null> => {
-    try {
-      const params = new URLSearchParams();
-      if (meterKey) params.set('meterKey', meterKey);
-      if (since) params.set('since', since);
-      const res = await fetchWithAuth(
-        `${API_BASE}/v0/management/quotas/${checkerId}/history?${params}`
-      );
-      if (!res.ok) throw new Error('Failed to fetch quota history');
-      return (await res.json()) as {
-        checkerId: string;
-        meterKey?: string;
-        since?: string;
-        history: any[];
-      };
-    } catch (e) {
-      console.error('API Error getQuotaHistory', e);
-      return null;
-    }
-  },
-
-  triggerQuotaCheck: async (checkerId: string): Promise<QuotaCheckerInfo | null> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/quotas/${checkerId}/check`, {
-        method: 'POST',
-      });
-      if (!res.ok) throw new Error('Failed to trigger quota check');
-      return normalizeQuotaCheckerInfo((await res.json()) as QuotaCheckerInfo);
-    } catch (e) {
-      console.error('API Error triggerQuotaCheck', e);
-      return null;
-    }
-  },
-
-  deleteAlias: async (aliasId: string): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/models/${encodeURIComponent(aliasId)}`,
-      {
-        method: 'DELETE',
-      }
-    );
-
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to delete alias');
-    }
-  },
-
-  deleteAllAliases: async (): Promise<void> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/models`, {
-      method: 'DELETE',
-    });
-
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to delete all aliases');
-    }
-  },
-
-  getOAuthProviders: async (): Promise<OAuthProviderInfo[]> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/oauth/providers`);
-    if (!res.ok) throw new Error('Failed to fetch OAuth providers');
-    const json = (await res.json()) as BackendResponse<OAuthProviderInfo[]>;
-    return json.data || [];
-  },
-
-  startOAuthSession: async (providerId: string, accountId: string): Promise<OAuthSession> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/oauth/sessions`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ providerId, accountId }),
-    });
-    if (!res.ok) {
-      const err = await res.json();
-      throw new Error(err.error || 'Failed to start OAuth session');
-    }
-    const json = (await res.json()) as { data: OAuthSession };
-    return json.data;
-  },
-
-  deleteOAuthCredentials: async (providerId: string, accountId: string): Promise<void> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/oauth/credentials`, {
-      method: 'DELETE',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ providerId, accountId }),
-    });
-    if (!res.ok) {
-      const err = await res.json();
-      throw new Error(err.error || 'Failed to delete OAuth credentials');
-    }
-  },
-
-  getOAuthCredentialStatus: async (
-    providerId: string,
-    accountId: string
-  ): Promise<OAuthCredentialStatus> => {
-    const query = new URLSearchParams({ providerId, accountId }).toString();
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/oauth/credentials/status?${query}`);
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to fetch OAuth credential status');
-    }
-    const json = (await res.json()) as { data: OAuthCredentialStatus };
-    return json.data;
-  },
-
-  getOAuthSession: async (sessionId: string): Promise<OAuthSession> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/oauth/sessions/${sessionId}`);
-    if (!res.ok) {
-      const err = await res.json();
-      throw new Error(err.error || 'Failed to fetch OAuth session');
-    }
-    const json = (await res.json()) as { data: OAuthSession };
-    return json.data;
-  },
-
-  submitOAuthPrompt: async (sessionId: string, value: string): Promise<OAuthSession> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/oauth/sessions/${sessionId}/prompt`,
-      {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ value }),
-      }
-    );
-    if (!res.ok) {
-      const err = await res.json();
-      throw new Error(err.error || 'Failed to submit OAuth prompt');
-    }
-    const json = (await res.json()) as { data: OAuthSession };
-    return json.data;
-  },
-
-  submitOAuthManualCode: async (sessionId: string, value: string): Promise<OAuthSession> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/oauth/sessions/${sessionId}/manual-code`,
-      {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ value }),
-      }
-    );
-    if (!res.ok) {
-      const err = await res.json();
-      throw new Error(err.error || 'Failed to submit OAuth code');
-    }
-    const json = (await res.json()) as { data: OAuthSession };
-    return json.data;
-  },
-
-  cancelOAuthSession: async (sessionId: string): Promise<OAuthSession> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/oauth/sessions/${sessionId}/cancel`,
-      {
-        method: 'POST',
-      }
-    );
-    if (!res.ok) {
-      const err = await res.json();
-      throw new Error(err.error || 'Failed to cancel OAuth session');
-    }
-    const json = (await res.json()) as { data: OAuthSession };
-    return json.data;
-  },
-
-  /**
-   * Search model metadata from an external catalog source.
-   * Used for autocomplete when assigning metadata to a model alias.
-   *
-   * @param source - "openrouter" | "models.dev" | "catwalk"
-   * @param query  - substring search (empty string = return all, up to limit)
-   * @param limit  - max results (default 50)
-   */
-  searchModelMetadata: async (
-    source: CatalogMetadataSource,
-    query?: string,
-    limit?: number
-  ): Promise<{ data: { id: string; name: string }[]; count: number }> => {
-    const params = new URLSearchParams({ source });
-    if (query) params.set('q', query);
-    if (limit !== undefined) params.set('limit', String(limit));
-    const res = await fetch(`${API_BASE}/v1/metadata/search?${params}`);
-    if (!res.ok) {
-      // 503 means the source isn't loaded yet — return empty gracefully
-      if (res.status === 503) return { data: [], count: 0 };
-      throw new Error(`Failed to search model metadata: ${res.statusText}`);
-    }
-    const result = (await res.json()) as {
-      data: { id: string; name: string }[];
-      count: number;
-    };
-    const data = dedupeById(result.data);
-    return { data, count: data.length };
-  },
-
-  /**
-   * Look up full catalog metadata for a specific model. Returns null when the
-   * source has not loaded (503) or the source_path is not found (404) so callers
-   * can gracefully fall back to leaving the form blank.
-   */
-  getModelMetadata: async (
-    source: CatalogMetadataSource,
-    sourcePath: string
-  ): Promise<NormalizedModelMetadata | null> => {
-    const params = new URLSearchParams({ source, source_path: sourcePath });
-    const res = await fetch(`${API_BASE}/v1/metadata/lookup?${params}`);
-    if (res.status === 404 || res.status === 503) return null;
-    if (!res.ok) {
-      throw new Error(`Failed to look up model metadata: ${res.statusText}`);
-    }
-    const json = (await res.json()) as { data: NormalizedModelMetadata };
-    return json.data;
-  },
-
-  refreshModelMetadata: async (): Promise<ModelMetadataRefreshResult> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/models/metadata/refresh`, {
-      method: 'POST',
-    });
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to refresh model metadata');
-    }
-    return (await res.json()) as ModelMetadataRefreshResult;
-  },
-
-  getPiProviders: async (): Promise<string[]> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/pi/providers`);
-    if (!res.ok) throw new Error('Failed to fetch pi providers');
-    const json = (await res.json()) as { data: string[] };
-    return json.data;
-  },
-
-  getPiModels: async (
-    provider: string,
-    q?: string
-  ): Promise<Array<{ id: string; name: string; api: string; custom: boolean }>> => {
-    const params = new URLSearchParams({ provider });
-    if (q) params.set('q', q);
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/pi/models?${params}`);
-    if (!res.ok) throw new Error('Failed to fetch pi models');
-    const json = (await res.json()) as {
-      data: Array<{ id: string; name: string; api: string; custom: boolean }>;
-    };
-    return dedupeById(json.data);
-  },
-
-  getOAuthProviderModels: async (
-    providerId: string,
-    accountId?: string
-  ): Promise<OAuthProviderModelsResult> => {
-    const query = new URLSearchParams({ providerId });
-    const trimmedAccountId = accountId?.trim();
-    if (trimmedAccountId) query.set('accountId', trimmedAccountId);
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/oauth/models?${query.toString()}`);
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to fetch OAuth provider models');
-    }
-    const json = (await res.json()) as {
-      data?: OAuthDiscoveredModel[];
-      source?: 'codex-backend' | 'catalog';
-      warning?: string;
-    };
-    return {
-      models: json.data || [],
-      source: json.source ?? 'catalog',
-      ...(json.warning ? { warning: json.warning } : {}),
-    };
-  },
-
-  getMcpServers: async (): Promise<Record<string, McpServer>> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/mcp-servers`);
-      if (!res.ok) throw new Error('Failed to fetch MCP servers');
-      return await res.json();
-    } catch (e) {
-      console.error('API Error getMcpServers', e);
-      return {};
-    }
-  },
-
-  saveMcpServer: async (serverName: string, server: McpServer): Promise<void> => {
-    try {
-      const res = await fetchWithAuth(
-        `${API_BASE}/v0/management/mcp-servers/${encodeURIComponent(serverName)}`,
-        {
-          method: 'PUT',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(server),
-        }
-      );
-      if (!res.ok) {
-        const err = await res.json().catch(() => ({}));
-        throw new Error(err.error || 'Failed to save MCP server');
-      }
-    } catch (e) {
-      console.error('API Error saveMcpServer', e);
-      throw e;
-    }
-  },
-
-  deleteMcpServer: async (serverName: string): Promise<void> => {
-    try {
-      const res = await fetchWithAuth(
-        `${API_BASE}/v0/management/mcp-servers/${encodeURIComponent(serverName)}`,
-        {
-          method: 'DELETE',
-        }
-      );
-      if (!res.ok) {
-        const err = await res.json().catch(() => ({}));
-        throw new Error(err.error || 'Failed to delete MCP server');
-      }
-    } catch (e) {
-      console.error('API Error deleteMcpServer', e);
-      throw e;
-    }
-  },
-
-  getMcpServerKeys: async (serverName: string): Promise<McpServerKey[]> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/mcp-servers/${encodeURIComponent(serverName)}/keys`,
-      { cache: 'no-store' }
-    );
-    if (!res.ok) throw new Error('Failed to fetch MCP server keys');
-    const data = (await res.json()) as { keys: McpServerKey[] };
-    return data.keys;
-  },
-
-  addMcpServerKey: async (serverName: string, key: string): Promise<McpServerKey> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/mcp-servers/${encodeURIComponent(serverName)}/keys`,
-      {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ key }),
-      }
-    );
-    if (!res.ok) {
-      const error = await res.json().catch(() => ({}));
-      throw new Error(error.error || 'Failed to add MCP server key');
-    }
-    return await res.json();
-  },
-
-  deleteMcpServerKey: async (serverName: string, keyId: number): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/mcp-servers/${encodeURIComponent(serverName)}/keys/${keyId}`,
-      { method: 'DELETE' }
-    );
-    if (!res.ok) throw new Error('Failed to delete MCP server key');
-  },
-
-  clearMcpServerKeyCooldown: async (serverName: string, keyId: number): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/mcp-servers/${encodeURIComponent(serverName)}/keys/${keyId}/clear-cooldown`,
-      { method: 'POST' }
-    );
-    if (!res.ok) throw new Error('Failed to clear MCP server key cooldown');
-  },
-
-  getMcpServerStatus: async (serverName: string): Promise<LocalMcpRuntimeStatus> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/mcp-servers/${encodeURIComponent(serverName)}/status`
-    );
-    if (!res.ok) throw new Error('Failed to fetch MCP server status');
-    return await res.json();
-  },
-
-  startMcpServer: async (serverName: string): Promise<LocalMcpRuntimeStatus> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/mcp-servers/${encodeURIComponent(serverName)}/start`,
-      { method: 'POST' }
-    );
-    if (!res.ok) throw new Error('Failed to start MCP server');
-    return await res.json();
-  },
-
-  stopMcpServer: async (serverName: string): Promise<LocalMcpRuntimeStatus> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/mcp-servers/${encodeURIComponent(serverName)}/stop`,
-      { method: 'POST' }
-    );
-    if (!res.ok) throw new Error('Failed to stop MCP server');
-    return await res.json();
-  },
-
-  restartMcpServer: async (serverName: string): Promise<LocalMcpRuntimeStatus> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/mcp-servers/${encodeURIComponent(serverName)}/restart`,
-      { method: 'POST' }
-    );
-    if (!res.ok) throw new Error('Failed to restart MCP server');
-    return await res.json();
-  },
-
-  getMcpLogs: async (
-    limit: number = 20,
-    offset: number = 0,
-    filters: { serverName?: string; apiKey?: string } = {}
-  ): Promise<{ data: McpLogRecord[]; total: number }> => {
-    try {
-      const params = new URLSearchParams({
-        limit: limit.toString(),
-        offset: offset.toString(),
-        ...(filters.serverName ? { serverName: filters.serverName } : {}),
-        ...(filters.apiKey ? { apiKey: filters.apiKey } : {}),
-      });
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/mcp-logs?${params}`);
-      if (!res.ok) throw new Error('Failed to fetch MCP logs');
-      return await res.json();
-    } catch (e) {
-      console.error('API Error getMcpLogs', e);
-      return { data: [], total: 0 };
-    }
-  },
-
-  getMcpOAuthClients: async (): Promise<McpOAuthClientRecord[]> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/mcp-oauth/clients`);
-      if (!res.ok) throw new Error('Failed to fetch MCP OAuth clients');
-      const body = (await res.json()) as { clients: McpOAuthClientRecord[] };
-      return body.clients || [];
-    } catch (e) {
-      console.error('API Error getMcpOAuthClients', e);
-      return [];
-    }
-  },
-
-  revokeMcpOAuthToken: async (tokenId: number): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/mcp-oauth/tokens/${encodeURIComponent(String(tokenId))}/revoke`,
-      { method: 'POST' }
-    );
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to revoke MCP OAuth token');
-    }
-  },
-
-  updateMcpOAuthClientStatus: async (
-    clientId: string,
-    status: 'active' | 'disabled'
-  ): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/mcp-oauth/clients/${encodeURIComponent(clientId)}`,
-      {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ status }),
-      }
-    );
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to update MCP OAuth client status');
-    }
-  },
-
-  deleteMcpOAuthClient: async (clientId: string): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/mcp-oauth/clients/${encodeURIComponent(clientId)}`,
-      { method: 'DELETE' }
-    );
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to delete MCP OAuth client');
-    }
-  },
-
-  revokeMcpOAuthClientTokens: async (clientId: string): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/mcp-oauth/clients/${encodeURIComponent(clientId)}/revoke`,
-      { method: 'POST' }
-    );
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to revoke MCP OAuth client tokens');
-    }
-  },
-
-  deleteMcpLog: async (requestId: string): Promise<boolean> => {
-    try {
-      const res = await fetchWithAuth(
-        `${API_BASE}/v0/management/mcp-logs/${encodeURIComponent(requestId)}`,
-        {
-          method: 'DELETE',
-        }
-      );
-      return res.ok;
-    } catch (e) {
-      console.error('API Error deleteMcpLog', e);
-      return false;
-    }
-  },
-
-  deleteAllMcpLogs: async (olderThanDays?: number): Promise<boolean> => {
-    try {
-      const params = olderThanDays != null ? `?olderThanDays=${olderThanDays}` : '';
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/mcp-logs${params}`, {
-        method: 'DELETE',
-      });
-      return res.ok;
-    } catch (e) {
-      console.error('API Error deleteAllMcpLogs', e);
-      return false;
-    }
-  },
-
-  // User Quota Management
-  getUserQuotas: async (): Promise<Record<string, UserQuota>> => {
-    try {
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/user-quotas`);
-      if (!res.ok) throw new Error('Failed to fetch user quotas');
-      return await res.json();
-    } catch (e) {
-      console.error('API Error getUserQuotas', e);
-      return {};
-    }
-  },
-
-  getUserQuota: async (name: string): Promise<UserQuota | null> => {
-    try {
-      const res = await fetchWithAuth(
-        `${API_BASE}/v0/management/user-quotas/${encodeURIComponent(name)}`
-      );
-      if (!res.ok) {
-        if (res.status === 404) return null;
-        throw new Error('Failed to fetch user quota');
-      }
-      return await res.json();
-    } catch (e) {
-      console.error('API Error getUserQuota', e);
-      return null;
-    }
-  },
-
-  saveUserQuota: async (name: string, quota: UserQuota): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/user-quotas/${encodeURIComponent(name)}`,
-      {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(quota),
-      }
-    );
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({ error: 'Unknown error' }));
-      throw new Error(err.error?.message || err.error || 'Failed to save quota');
-    }
-  },
-
-  updateUserQuota: async (name: string, updates: Partial<UserQuota>): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/user-quotas/${encodeURIComponent(name)}`,
-      {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(updates),
-      }
-    );
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({ error: 'Unknown error' }));
-      throw new Error(err.error?.message || err.error || 'Failed to update quota');
-    }
-  },
-
-  deleteUserQuota: async (name: string): Promise<void> => {
-    const res = await fetchWithAuth(
-      `${API_BASE}/v0/management/user-quotas/${encodeURIComponent(name)}`,
-      {
-        method: 'DELETE',
-      }
-    );
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({ error: 'Unknown error' }));
-      throw new Error(err.error?.message || err.error || 'Failed to delete quota');
-    }
-  },
-
-  getQuotaStatus: async (
-    key: string
-  ): Promise<{
-    key: string;
-    quotas: QuotaStatusEntry[];
-    // Legacy shim fields, derived from the most-constrained entry in `quotas`.
-    quota_name: string | null;
-    allowed: boolean;
-    current_usage: number;
-    limit: number | null;
-    remaining: number | null;
-    resets_at: string | null;
-  } | null> => {
-    try {
-      const res = await fetchWithAuth(
-        `${API_BASE}/v0/management/quota/status/${encodeURIComponent(key)}`
-      );
-      if (!res.ok) {
-        if (res.status === 404) return null;
-        throw new Error('Failed to fetch quota status');
-      }
-      return await res.json();
-    } catch (e) {
-      console.error('API Error getQuotaStatus', e);
-      return null;
-    }
-  },
-
-  /** Reset usage for a key's quota(s). Omit `quota` to clear every quota
-   * currently attached to the key. */
-  clearQuota: async (key: string, quota?: string): Promise<void> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/quota/clear`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ key, ...(quota ? { quota } : {}) }),
-    });
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({ error: 'Unknown error' }));
-      throw new Error(err.error?.message || err.error || 'Failed to clear quota');
-    }
-  },
-
-  /**
-   * Repair a quota bucket by recomputing it from request_usage instead of
-   * trusting the (potentially drifted) counter. Rejected (400, `reason:
-   * 'unsupported_quota_type'`) for rolling requests/tokens defs, which are
-   * inherently leaky (see backend `QuotaEnforcer.recomputeQuota`) — the
-   * thrown error carries `.reason` so callers can special-case that message.
-   */
-  recomputeQuota: async (
-    key: string,
-    quota: string
-  ): Promise<{
-    success: true;
-    key: string;
-    quota: string;
-    usage: number;
-    windowStartMs: number;
-    message: string;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/quota/recompute`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ key, quota }),
-    });
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({ error: 'Unknown error' }));
-      const message = err.error?.message || err.error || 'Failed to recompute quota';
-      const wrapped = new Error(message) as Error & { reason?: string };
-      wrapped.reason = err.reason;
-      throw wrapped;
-    }
-    return res.json();
-  },
-
-  /**
-   * Fetches concurrency data from the backend.
-   *
-   * Calls GET /v0/management/concurrency with mode, timeRange, and groupBy query parameters.
-   * - mode='live': returns current in-flight snapshots.
-   * - mode='timeline': returns bucketed historical counts.
-   *
-   * On failure, logs the error and returns an empty array so the UI degrades
-   * gracefully (shows an empty chart rather than crashing).
-   *
-   * @param timeRange - How far back to look: 'hour' (default), 'day', 'week', or 'month'
-   * @param groupBy - Dimension to group by: 'provider' (default) or 'model'
-   * @returns Array of {@link ConcurrencyData} entries, or an empty array on error
-   */
-  getConcurrencyData: async (
-    timeRange: 'hour' | 'day' | 'week' | 'month' | 'custom' = 'hour',
-    mode: 'live' | 'timeline' = 'live',
-    groupBy: 'provider' | 'model' = 'provider',
-    startDate?: string,
-    endDate?: string
-  ): Promise<ConcurrencyData[]> => {
-    try {
-      const params = new URLSearchParams();
-      params.set('timeRange', timeRange);
-      params.set('mode', mode);
-      params.set('groupBy', groupBy);
-
-      if (timeRange === 'custom' && startDate && endDate) {
-        params.set('startDate', startDate);
-        params.set('endDate', endDate);
-      }
-
-      const res = await fetchWithAuth(`${API_BASE}/v0/management/concurrency?${params.toString()}`);
-      if (!res.ok) throw new Error('Failed to fetch concurrency data');
-      const data = (await res.json()) as { data: ConcurrencyData[] };
-      return data.data || [];
-    } catch (e) {
-      console.error('API Error getConcurrencyData', e);
-      return [];
-    }
-  },
-
-  /**
-   * Fetches models from a provider API via server-side proxy.
-   * This bypasses CORS restrictions by routing the request through the backend.
-   *
-   * @param url - The provider's models endpoint URL
-   * @param apiKey - Optional API key for authentication
-   * @returns Array of models in OpenAI format { data: [...] }
-   */
-  fetchProviderModels: async (
-    url: string,
-    apiKey?: string
-  ): Promise<{
-    data: Array<{ id: string; object?: string; created?: number; owned_by?: string }>;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/providers/fetch-models`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ url, apiKey }),
-    });
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({ error: { message: 'Unknown error' } }));
-      // Handle API.md error format: { error: { message, type, code, details } }
-      const errorMessage =
-        err.error?.message || err.error || err.details || 'Failed to fetch models';
-      throw new Error(errorMessage);
-    }
-    return res.json();
-  },
-
-  // ─── Self-service (limited user operating on their own key) ──────────
-
-  /**
-   * Fetch metadata about the current principal's key (name, allowed providers /
-   * models, quota, comment, trace state).
-   */
-  getSelfMe: async (): Promise<{
-    role: 'admin' | 'limited';
-    keyName?: string;
-    allowedProviders?: string[];
-    allowedModels?: string[];
-    allowRawPassthrough?: boolean;
-    quotaNames?: string[];
-    quotaName?: string | null;
-    comment?: string | null;
-    traceEnabled?: boolean;
-    traceEnabledGlobal?: boolean;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/self/me`);
-    if (!res.ok) throw new Error('Failed to fetch self info');
-    return res.json();
-  },
-
-  /** Rotate the current principal's secret. Returns the new plaintext once. */
-  rotateSelfSecret: async (): Promise<{ keyName: string; secret: string; message: string }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/self/rotate`, {
-      method: 'POST',
-    });
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({ error: 'Rotation failed' }));
-      throw new Error(err.error?.message || err.error || 'Rotation failed');
-    }
-    return res.json();
-  },
-
-  /** Update the current principal's key comment. */
-  updateSelfComment: async (
-    comment: string | null
-  ): Promise<{ success: boolean; keyName: string; comment: string | null }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/self/comment`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ comment }),
-    });
-    if (!res.ok) throw new Error('Failed to update comment');
-    return res.json();
-  },
-
-  /** Enable or disable trace capture for the current principal's key only. */
-  toggleSelfDebug: async (
-    enabled: boolean
-  ): Promise<{ keyName: string; enabled: boolean; enabledGlobal: boolean }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/self/debug/toggle`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ enabled }),
-    });
-    if (!res.ok) throw new Error('Failed to toggle trace');
-    return res.json();
-  },
-
-  /**
-   * Fetch the quota status for the current principal's key.
-   *
-   * Returns a payload with `quotaName: null` if the key has no quota
-   * assigned (so callers can render "No quota" instead of handling a
-   * distinct error case).
-   */
-  getSelfQuota: async (): Promise<{
-    key: string;
-    quotas: QuotaStatusEntry[];
-    // Legacy shim fields, derived from the most-constrained entry in `quotas`.
-    quotaName: string | null;
-    allowed: boolean;
-    currentUsage: number;
-    limit: number | null;
-    remaining: number | null;
-    resetsAt: string | null;
-    limitType: 'requests' | 'tokens' | 'cost' | null;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/self/quota`);
-    if (!res.ok) throw new Error('Failed to fetch quota status');
-    return res.json();
-  },
-
-  /** Export a config-only backup as JSON. */
-  createBackup: async (): Promise<Blob> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/backup`);
-    if (!res.ok) throw new Error('Failed to create backup');
-    return res.blob();
-  },
-
-  /** Export a full backup (config + operational data) as a .tar.gz archive. */
-  createFullBackup: async (): Promise<Blob> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/backup?full=true`);
-    if (!res.ok) throw new Error('Failed to create full backup');
-    return res.blob();
-  },
-
-  /** Restore from a config-only JSON backup. */
-  restoreBackup: async (
-    data: object
-  ): Promise<{
-    success: boolean;
-    restored: Record<string, number>;
-    message: string;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/restore`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(data),
-    });
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({ error: 'Restore failed' }));
-      throw new Error(err.error || 'Restore failed');
-    }
-    return res.json();
-  },
-
-  /** Restore from a full .tar.gz backup archive. */
-  restoreFullBackup: async (
-    file: File
-  ): Promise<{
-    success: boolean;
-    restored: Record<string, number>;
-    message: string;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/restore`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/octet-stream' },
-      body: await file.arrayBuffer(),
-    } as RequestInit);
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({ error: 'Restore failed' }));
-      throw new Error(err.error || 'Restore failed');
-    }
-    return res.json();
-  },
-
-  /** Reset all request logs, error logs, and debug trace logs. */
-  resetLogs: async (): Promise<{ success: boolean; message: string }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/logs/reset`, {
-      method: 'DELETE',
-    });
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({ error: 'Reset logs failed' }));
-      throw new Error(err.error || 'Reset logs failed');
-    }
-    return res.json();
-  },
-
-  getSystemSettings: async (): Promise<Record<string, unknown>> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/system-settings`);
-    if (!res.ok) throw new Error('Failed to fetch system settings');
-    return res.json();
-  },
-
-  patchSystemSettings: async (updates: Record<string, unknown>): Promise<void> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/system-settings`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(updates),
-    });
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      throw new Error(err.error || 'Failed to update system settings');
-    }
-  },
-
-  // ─── Failover Settings ────────────────────────────────────────────
-
-  /** Fetch current failover policy. */
-  getFailoverPolicy: async (): Promise<{
-    enabled: boolean;
-    retryableStatusCodes: number[];
-    retryableErrors: string[];
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/failover`);
-    if (!res.ok) throw new Error('Failed to fetch failover policy');
-    return res.json();
-  },
-
-  /** Patch failover policy fields. */
-  patchFailoverPolicy: async (updates: {
-    enabled?: boolean;
-    retryableStatusCodes?: number[];
-    retryableErrors?: string[];
-  }): Promise<{
-    enabled: boolean;
-    retryableStatusCodes: number[];
-    retryableErrors: string[];
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/failover`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(updates),
-    });
-    if (!res.ok) throw new Error('Failed to update failover policy');
-    return res.json();
-  },
-
-  // ─── Capture Trace on Error ─────────────────────────────────────────
-
-  /** Fetch whether traces are captured on error even when debug is off. */
-  getCaptureTraceOnError: async (): Promise<{ enabled: boolean }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/capture-trace-on-error`);
-    if (!res.ok) throw new Error('Failed to fetch capture-trace-on-error setting');
-    return res.json();
-  },
-
-  /** Toggle capturing traces on error even when debug is off. */
-  setCaptureTraceOnError: async (enabled: boolean): Promise<{ enabled: boolean }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/capture-trace-on-error`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ enabled }),
-    });
-    if (!res.ok) throw new Error('Failed to update capture-trace-on-error setting');
-    return res.json();
-  },
-
-  // ─── Cooldown Settings ──────────────────────────────────────────────
-
-  /** Fetch current cooldown policy. */
-  getCooldownPolicy: async (): Promise<{
-    initialMinutes: number;
-    maxMinutes: number;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/cooldown`);
-    if (!res.ok) throw new Error('Failed to fetch cooldown policy');
-    return res.json();
-  },
-
-  /** Patch cooldown policy fields. */
-  patchCooldownPolicy: async (updates: {
-    initialMinutes?: number;
-    maxMinutes?: number;
-  }): Promise<{
-    initialMinutes: number;
-    maxMinutes: number;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/cooldown`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(updates),
-    });
-    if (!res.ok) throw new Error('Failed to update cooldown policy');
-    return res.json();
-  },
-
-  // ─── Trusted Proxies ───────────────────────────────────────────────
-
-  /** Fetch the trusted-proxy allowlist (IPs/CIDRs whose forwarding headers are honored). */
-  getTrustedProxies: async (): Promise<{ trustedProxies: string[] }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/trusted-proxies`);
-    if (!res.ok) throw new Error('Failed to fetch trusted proxies');
-    return res.json();
-  },
-
-  /** Replace the trusted-proxy allowlist. */
-  patchTrustedProxies: async (trustedProxies: string[]): Promise<{ trustedProxies: string[] }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/trusted-proxies`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ trustedProxies }),
-    });
-    if (!res.ok) {
-      const err = await res.json().catch(() => ({}));
-      const detail =
-        Array.isArray(err.details) && err.details[0]?.message ? `: ${err.details[0].message}` : '';
-      throw new Error(`${err.error || 'Failed to update trusted proxies'}${detail}`);
-    }
-    return res.json();
-  },
-
-  // ─── Exploration Rate Settings ─────────────────────────────────────
-
-  /** Fetch current exploration rate settings. */
-  getExplorationRates: async (): Promise<{
-    performanceExplorationRate: number;
-    latencyExplorationRate: number;
-    e2ePerformanceExplorationRate: number;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/exploration-rate`);
-    if (!res.ok) throw new Error('Failed to fetch exploration rate settings');
-    return res.json();
-  },
-
-  /** Patch exploration rate settings. */
-  patchExplorationRates: async (updates: {
-    performanceExplorationRate?: number;
-    latencyExplorationRate?: number;
-    e2ePerformanceExplorationRate?: number;
-  }): Promise<{
-    performanceExplorationRate: number;
-    latencyExplorationRate: number;
-    e2ePerformanceExplorationRate: number;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/exploration-rate`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(updates),
-    });
-    if (!res.ok) throw new Error('Failed to update exploration rate settings');
-    return res.json();
-  },
-
-  // ─── Background Exploration Settings ──────────────────────────
-
-  /** Fetch current background exploration settings. */
-  getBackgroundExploration: async (): Promise<{
-    enabled: boolean;
-    stalenessThresholdSeconds: number;
-    workerConcurrency: number;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/background-exploration`);
-    if (!res.ok) throw new Error('Failed to fetch background exploration settings');
-    return res.json();
-  },
-
-  /** Patch background exploration settings. */
-  patchBackgroundExploration: async (updates: {
-    enabled?: boolean;
-    stalenessThresholdSeconds?: number;
-    workerConcurrency?: number;
-  }): Promise<{
-    enabled: boolean;
-    stalenessThresholdSeconds: number;
-    workerConcurrency: number;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/background-exploration`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(updates),
-    });
-    if (!res.ok) throw new Error('Failed to update background exploration settings');
-    return res.json();
-  },
-
-  // ─── Timeout Settings ───────────────────────────────────────────
-
-  /** Fetch current timeout settings. */
-  getTimeoutConfig: async (): Promise<{ defaultSeconds: number }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/timeout`);
-    if (!res.ok) throw new Error('Failed to fetch timeout settings');
-    return res.json();
-  },
-
-  /** Patch timeout settings. */
-  patchTimeoutConfig: async (updates: {
-    defaultSeconds?: number;
-  }): Promise<{ defaultSeconds: number }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/timeout`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(updates),
-    });
-    if (!res.ok) throw new Error('Failed to update timeout settings');
-    return res.json();
-  },
-
-  // ─── Compaction Settings ─────────────────────────────────────────
-
-  /** Fetch current compaction settings. */
-  getCompactionConfig: async (): Promise<CompactionSettings> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/compaction`);
-    if (!res.ok) throw new Error('Failed to fetch compaction settings');
-    return res.json();
-  },
-
-  /** Patch compaction settings. */
-  patchCompactionConfig: async (updates: CompactionSettings): Promise<CompactionSettings> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/compaction`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(updates, (_key, value) => (value === undefined ? null : value)),
-    });
-    if (!res.ok) throw new Error('Failed to update compaction settings');
-    return res.json();
-  },
-
-  // ─── MCP Server Enabled ───────────────────────────────────────────
-
-  /** Fetch current MCP server enabled state. */
-  getMcpEnabled: async (): Promise<{ enabled: boolean }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/mcp-enabled`);
-    if (!res.ok) throw new Error('Failed to fetch MCP enabled state');
-    return res.json();
-  },
-
-  /** Enable or disable the MCP server. */
-  patchMcpEnabled: async (enabled: boolean): Promise<{ enabled: boolean }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/mcp-enabled`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ enabled }),
-    });
-    if (!res.ok) throw new Error('Failed to update MCP enabled state');
-    return res.json();
-  },
-
-  // ─── Stall Detection Settings ────────────────────────────────────
-
-  /** Fetch current stall detection settings. */
-  getStallConfig: async (): Promise<{
-    ttfbSeconds: number | null;
-    ttfbBytes: number;
-    minBytesPerSecond: number | null;
-    windowSeconds: number;
-    gracePeriodSeconds: number;
-    stallCooldown: boolean;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/stall`);
-    if (!res.ok) throw new Error('Failed to fetch stall detection settings');
-    return res.json();
-  },
-
-  /** Patch stall detection settings. */
-  patchStallConfig: async (updates: {
-    ttfbSeconds?: number | null;
-    ttfbBytes?: number;
-    minBytesPerSecond?: number | null;
-    windowSeconds?: number;
-    gracePeriodSeconds?: number;
-    stallCooldown?: boolean;
-  }): Promise<{
-    ttfbSeconds: number | null;
-    ttfbBytes: number;
-    minBytesPerSecond: number | null;
-    windowSeconds: number;
-    gracePeriodSeconds: number;
-    stallCooldown: boolean;
-  }> => {
-    const res = await fetchWithAuth(`${API_BASE}/v0/management/config/stall`, {
-      method: 'PATCH',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(updates),
-    });
-    if (!res.ok) throw new Error('Failed to update stall detection settings');
-    return res.json();
-  },
-};
+// Domain Types
+export * from '../types/usage';
+export * from '../types/mcp';
+export * from '../types/aliases';
+export * from '../types/settings';
+export * from '../types/quota';
+
+// API Client & Methods
+export * from './api/index';
